+++ fwb/app/(auth)/layout.tsx
@@ -4,7 +4,7 @@ const AuthLayout = ({
   children: React.ReactNode;
 }) => {
   return (
-    <div className="flex items-center justify-center h-full">
+    <div>
       {children}
     </div>
   );
