+++ fwb/app/(fre)/fre1/page.css
@@ -1,6 +1,6 @@
 @import url(https://fonts.googleapis.com/css2?family=Cormorant:wght@300;400&family=Noto+Sans&family=Nunito+Sans&family=Outfit&family=Playfair+Display&family=Poppins:wght@300&family=Urbanist&family=Yantramanav:wght@100;300;400&display=swap);
 
-.big {
+.pageContent {
   /* max-width: 1440px;
   max-height: 931px; */
   display: flex;
@@ -8,7 +8,7 @@
   height: 100vh;
 }
 
-.spacing {
+.middleSpacing {
   flex-shrink: 0;
   padding-left: 142px;
   padding-right: 143px;
@@ -32,7 +32,7 @@
   margin-top: 103px;
 }
 
-.welcometitle {
+.mainHeader {
   color: var(--neutral-000, #fff);
   text-align: center;
   /* Heading/H2 */
@@ -45,44 +45,41 @@
   letter-spacing: 0.1rem;
 }
 
-.letscreate {
-    color: var(--neutral-000, #FFF);
-    text-align: center;
-    /* Heading/H5 */
-    font-family: Urbanist;
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 125%; /* 22.5px */
-    margin-top: 8px;
-    margin-bottom: 32px;
+.subtext {
+  color: var(--neutral-000, #FFF);
+  text-align: center;
+  /* Heading/H5 */
+  font-family: Urbanist;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 125%; /* 22.5px */
+  margin-top: 8px;
+  margin-bottom: 32px;
 }
 
-/*.profilepics {
-  display: flex;
-  justify-content: center;
-  margin-top: 32px;
-}*/
+.image {
+  border-radius: 50%;
+  background-image: url('/profilepics/PNG/WomanOne.png');
+  background-position: center;
+}
 
-.next {
-  border-radius: 30px;
-  background: var(--yellow-300, #f6ff82);
-  display: flex;
-  width: 420px;
-  height: 48px;
-  padding: 10px 24px;
-  justify-content: center;
-  gap: 8px;
-  color: var(--purple-300, #8e94e9);
-  text-align: center;
-  /* Label/24/bold */
+input[type="file"]::file-selector-button {
+  background-color: #f6ff82;
+  border: #f6ff82;
+  border-radius: 100px;
+  color:#8e94e9;
+}
+
+.addPhoto {
+  color: var(--neutral-000, #FFF);
+  /* Label/16/bold */
   font-family: Urbanist;
-  font-size: 20px;
+  font-size: 16px;
   font-style: normal;
   font-weight: 700;
-  line-height: 125%; /* 30px */
-  letter-spacing: 0.4px;
-  margin-bottom: 12px;
+  line-height: 125%; /* 20px */
+  letter-spacing: 0.32px;
 }
 
 .inputusername {
@@ -108,28 +105,25 @@ input::placeholder {
   opacity: 0.3;
 }
 
-.image {
-  border-radius: 50%;
-  background-image: url('/profilepics/PNG/WomanOne.png');
-  background-position: center;
-}
-
-input[type="file"]::file-selector-button {
-  background-color: #f6ff82;
-  border: #f6ff82;
-  border-radius: 100px;
-  color:#8e94e9;
-}
-
-.addPhoto {
-  color: var(--neutral-000, #FFF);
-  /* Label/16/bold */
+.next {
+  border-radius: 30px;
+  background: var(--yellow-300, #f6ff82);
+  display: flex;
+  width: 420px;
+  height: 48px;
+  padding: 10px 24px;
+  justify-content: center;
+  gap: 8px;
+  color: var(--purple-300, #8e94e9);
+  text-align: center;
+  /* Label/24/bold */
   font-family: Urbanist;
-  font-size: 16px;
+  font-size: 20px;
   font-style: normal;
   font-weight: 700;
-  line-height: 125%; /* 20px */
-  letter-spacing: 0.32px;
+  line-height: 125%; /* 30px */
+  letter-spacing: 0.4px;
+  margin-bottom: 12px;
 }
 
 .updatePhoto {
