+++ fwb/app/(fre)/fre2/page.css
@@ -1,6 +1,6 @@
 @import url(https://fonts.googleapis.com/css2?family=Cormorant:wght@300;400&family=Noto+Sans&family=Nunito+Sans&family=Outfit&family=Playfair+Display&family=Poppins:wght@300&family=Urbanist&family=Yantramanav:wght@100;300;400&display=swap);
 
-.big {
+.pageContent {
   /* max-width: 1440px;
   max-height: 931px; */
   display: flex;
@@ -8,19 +8,19 @@
   height: 100vh;
 }
 
-.spacing {
+.middleSpacing {
   flex-shrink: 0;
   padding-left: 142px;
   padding-right: 150px;
 }
 
-.illustrationFive {
+.illustrationThree {
   margin-left: 108px;
   height: 100%;
   flex-shrink: 0;
 }
 
-.illustrationSix {
+.illustrationFour {
   margin-right: 120px;
   flex-shrink: 0;
 }
@@ -31,7 +31,7 @@
   margin-top: 103px;
 }
 
-.largeheader {
+.mainHeader {
   color: var(--neutral-000, #fff);
   text-align: center;
   /* Heading/H2 */
@@ -40,34 +40,34 @@
   font-style: normal;
   font-weight: 600;
   line-height: 110%; /* 44px */
-  margin-top: 135px;
+  margin-top: 131px;
   margin-bottom:16px;
   letter-spacing: 0.1rem;
 }
 
-.lowertext {
-    color: var(--neutral-000, #FFF);
-    text-align: center;
-    /* Heading/H5 */
-    font-family: Urbanist;
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 125%; /* 22.5px */
-    margin-top: 8px;
+.subtextOne {
+  color: var(--neutral-000, #FFF);
+  text-align: center;
+  /* Heading/H5 */
+  font-family: Urbanist;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 125%; /* 22.5px */
+  margin-top: 8px;
 }
 
-.or {
+.subtextTwo {
   color: var(--neutral-000, #FFF);
-    text-align: center;
-    /* Heading/H5 */
-    font-family: Urbanist;
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 125%; /* 22.5px */
-    margin-top: 24px;
-    margin-bottom:24px
+  text-align: center;
+  /* Heading/H5 */
+  font-family: Urbanist;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 125%; /* 22.5px */
+  margin-top: 8px;
+  margin-bottom: 72px;
 }
 
 .next {
@@ -92,7 +92,7 @@
   margin-bottom: 8px;
 }
 
-.inputfriends {
+.inputCompany {
   border-radius: 100px;
   background: var(--neutral-000, #fff);
   display: flex;
@@ -101,7 +101,7 @@
   padding: 8px 8px 8px 24px;
   gap: 8px;
   margin-bottom: 13.53px;
-  margin-top: 72px;
+  margin-top: 32px;
 }
 
 input::placeholder {
@@ -115,7 +115,7 @@ input::placeholder {
   opacity: 0.3;
 }
 
-.linkedinverify {
+.linkedInVerify {
   border-radius: 30px;
   background: #8e94e9;
   border: 2px solid #FFF;
@@ -140,23 +140,22 @@ input::placeholder {
 
 .help{
   color: #FFF;
-
-font-family: Urbanist;
-font-size: 16px;
-font-style: normal;
-font-weight: 500;
-line-height: 150%; 
+  font-family: Urbanist;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 150%; 
 }
 
 .helpEmail{
   color: var(--neutral-000, #FFF);
-/* Label/16/bold */
-font-family: Urbanist;
-font-size: 16px;
-font-style: normal;
-font-weight: 700;
-line-height: 125%; /* 20px */
-letter-spacing: 0.32px;
+  /* Label/16/bold */
+  font-family: Urbanist;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 125%; /* 20px */
+  letter-spacing: 0.32px;
 }
 
 .skip {
