--- fwb/app/(fre)/testing/page.css
@@ -1,24 +1,0 @@
-@import url(https://fonts.googleapis.com/css2?family=Cormorant:wght@300;400&family=Noto+Sans&family=Nunito+Sans&family=Outfit&family=Playfair+Display&family=Poppins:wght@300&family=Urbanist&family=Yantramanav:wght@100;300;400&display=swap);
-
-.image {
-  border-radius: 50%;
-}
-
-
-.addPhoto {
-  color: var(--neutral-000, #FFF);
-  /* Label/16/bold */
-  font-family: Urbanist;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 125%; /* 20px */
-  letter-spacing: 0.32px;
-  }
-
-input[type="file"]::file-selector-button {
-  background-color: #8e94e9;
-  border: #8e94e9;
-}
-
-
