--- fwb/app/(fre)/uf3/page.tsx
@@ -1,99 +1,0 @@
-"use client";
-
-import './page.css';
-import IllustrationFive from '@/components/ui/fre/IllustrationFive';
-import IllustrationSix from '@/components/ui/fre/IllustrationSix';
-import Link from "next/link";
-
-export default function UserFlowPage() {
-
-  const handlelink1 = () => {
-    window.open('https://www.whatsapp.com/')
-  }
-
-  const handlelink2 = () => {
-    window.open('https://www.instagram.com/')
-  }
-
-  const handlelink3 = () => {
-    window.open('https://discord.com/')
-  }
-
-
-  return (
-    <div className="big">
-      <IllustrationFive />
-      <div className="spacing">
-        <div className="flex-col justify-center">
-          <div className="progresscircles">
-          <svg xmlns="http://www.w3.org/2000/svg" width="56" height="8" viewBox="0 0 56 8" fill="none">
-            <circle cx="4" cy="4" r="4" fill="#ADB4D2"/>
-            <circle cx="28" cy="4" r="4" fill="#ADB4D2"/>
-            <circle cx="52" cy="4" r="4" fill="#F6FF82"/>
-          </svg>
-          </div>
-            <h2 className="largeheader">Share with Your Friends!</h2>
-            <h5 className="lowertext">Spread the love and be the wingman to someone else's wallet!</h5>
-          <form className="flex justify-center">
-            <input
-              type="username"
-              className="inputfriends"
-              placeholder="Tag Friends!"
-            />
-          </form>
-
-          <h5 className='or'>Or</h5>
-
-          <div className='flex justify-center items-center space-x-4'>
-              <button onClick={handlelink1}>
-                  <svg xmlns="http://www.w3.org/2000/svg" width="46" height="46" viewBox="0 0 46 46" fill="none">
-                    <rect x="0.88228" y="1.17854" width="43.8029" height="43.8029" rx="21.9015" stroke="white" stroke-width="1.413"/>
-                    <path d="M34.2878 16.378C30.369 10.3029 22.344 8.50293 16.1377 12.2717C10.0814 16.0405 8.13144 24.253 12.0502 30.3092L12.3689 30.7967L11.0564 35.7093L15.969 34.3968L16.4565 34.7155C18.5752 35.8593 20.8815 36.5155 23.169 36.5155C25.6252 36.5155 28.0815 35.8593 30.2002 34.5468C36.2565 30.6092 38.0565 22.5655 34.2878 16.3405V16.378ZM30.8565 29.0155C30.2002 29.9905 29.3752 30.6467 28.2315 30.8155C27.5752 30.8155 26.7502 31.1342 23.4877 29.8405C20.7127 28.528 18.4065 26.3905 16.7752 23.9342C15.8002 22.7905 15.294 21.3092 15.144 19.828C15.144 18.5155 15.6315 17.3717 16.4565 16.5467C16.7752 16.228 17.1127 16.0592 17.4315 16.0592H18.2565C18.5752 16.0592 18.9127 16.0592 19.0815 16.7155C19.4002 17.5405 20.2252 19.5092 20.2252 19.678C20.394 19.8467 20.319 21.103 19.569 21.8155C19.1565 22.2842 19.0815 22.303 19.2502 22.6405C19.9065 23.6155 20.7315 24.6092 21.5377 25.4342C22.5127 26.2592 23.5065 26.9155 24.6502 27.403C24.969 27.5717 25.3065 27.5717 25.4752 27.2342C25.644 26.9155 26.4502 26.0905 26.7877 25.753C27.1065 25.4342 27.2752 25.4342 27.6127 25.5842L30.2377 26.8967C30.5565 27.0655 30.894 27.2155 31.0627 27.3842C31.2315 27.8717 31.2315 28.528 30.894 29.0155H30.8565Z" fill="white"/>
-                  </svg>
-                  </button>
-            <button onClick={handlelink2}>
-              <svg xmlns="http://www.w3.org/2000/svg" width="46" height="46" viewBox="0 0 46 46" fill="none">
-                <rect x="1.0981" y="1.17854" width="43.8029" height="43.8029" rx="21.9015" stroke="white" stroke-width="1.413"/>
-                <g clip-path="url(#clip0_1223_8574)">
-                <path d="M23.1088 18.5968C21.8909 18.5968 20.7229 19.0806 19.8616 19.9418C19.0004 20.8031 18.5166 21.9711 18.5166 23.189C18.5166 24.407 19.0004 25.575 19.8616 26.4362C20.7229 27.2975 21.8909 27.7813 23.1088 27.7813C24.3268 27.7813 25.4948 27.2975 26.356 26.4362C27.2173 25.575 27.7011 24.407 27.7011 23.189C27.7011 21.9711 27.2173 20.8031 26.356 19.9418C25.4948 19.0806 24.3268 18.5968 23.1088 18.5968Z" fill="white"/>
-                <path fill-rule="evenodd" clip-rule="evenodd" d="M15.7187 10.5879C20.6303 10.0435 25.587 10.0435 30.4986 10.5879C33.1819 10.8874 35.3452 12.9999 35.6603 15.693C36.2428 20.6734 36.2428 25.7046 35.6603 30.6849C35.3452 33.3781 33.1819 35.4905 30.5001 35.7915C25.588 36.3361 20.6308 36.3361 15.7187 35.7915C13.0354 35.4905 10.8721 33.3781 10.557 30.6864C9.97448 25.7056 9.97448 20.6738 10.557 15.693C10.8721 12.9999 13.0354 10.8874 15.7187 10.5879ZM30.1737 14.711C29.7989 14.711 29.4395 14.8599 29.1745 15.1249C28.9095 15.3899 28.7607 15.7493 28.7607 16.124C28.7607 16.4988 28.9095 16.8582 29.1745 17.1231C29.4395 17.3881 29.7989 17.537 30.1737 17.537C30.5484 17.537 30.9078 17.3881 31.1728 17.1231C31.4378 16.8582 31.5867 16.4988 31.5867 16.124C31.5867 15.7493 31.4378 15.3899 31.1728 15.1249C30.9078 14.8599 30.5484 14.711 30.1737 14.711ZM16.3969 23.189C16.3969 21.4089 17.1041 19.7018 18.3628 18.4431C19.6215 17.1844 21.3286 16.4773 23.1087 16.4773C24.8887 16.4773 26.5959 17.1844 27.8546 18.4431C29.1133 19.7018 29.8204 21.4089 29.8204 23.189C29.8204 24.9691 29.1133 26.6762 27.8546 27.9349C26.5959 29.1936 24.8887 29.9007 23.1087 29.9007C21.3286 29.9007 19.6215 29.1936 18.3628 27.9349C17.1041 26.6762 16.3969 24.9691 16.3969 23.189Z" fill="white"/>
-                </g>
-                <defs>
-                <clipPath id="clip0_1223_8574">
-                <rect width="25.4339" height="25.4339" fill="white" transform="translate(10.3916 10.472)"/>
-                </clipPath>
-                </defs>
-              </svg>
-            </button>
-            <button onClick={handlelink3}>
-              <svg xmlns="http://www.w3.org/2000/svg" width="46" height="46" viewBox="0 0 46 46" fill="none">
-                <rect x="1.31392" y="1.17854" width="43.8029" height="43.8029" rx="21.9015" stroke="white" stroke-width="1.413"/>
-                <path d="M32.4356 15.2299C30.7892 14.4596 29.0288 13.8998 27.1884 13.5811C26.9623 13.9897 26.6983 14.5393 26.5162 14.9766C24.5598 14.6824 22.6213 14.6824 20.7009 14.9766C20.5188 14.5393 20.2488 13.9897 20.0207 13.5811C18.1783 13.8998 16.4159 14.4617 14.7695 15.234C11.4487 20.2522 10.5485 25.1456 10.9986 29.9696C13.2011 31.6144 15.3356 32.6136 17.4341 33.2674C17.9522 32.5543 18.4143 31.7963 18.8124 30.9974C18.0543 30.7093 17.3281 30.3538 16.642 29.941C16.824 29.8062 17.002 29.6652 17.1741 29.5201C21.359 31.4775 25.9061 31.4775 30.041 29.5201C30.2151 29.6652 30.3931 29.8062 30.5732 29.941C29.885 30.3558 29.1568 30.7113 28.3987 30.9994C28.7968 31.7963 29.2569 32.5564 29.777 33.2694C31.8775 32.6156 34.014 31.6165 36.2165 29.9696C36.7446 24.3774 35.3143 19.5288 32.4356 15.2299ZM19.3826 27.0029C18.1263 27.0029 17.096 25.8301 17.096 24.4019C17.096 22.9737 18.1043 21.7989 19.3826 21.7989C20.6609 21.7989 21.6911 22.9716 21.6691 24.4019C21.6711 25.8301 20.6609 27.0029 19.3826 27.0029ZM27.8325 27.0029C26.5763 27.0029 25.546 25.8301 25.546 24.4019C25.546 22.9737 26.5542 21.7989 27.8325 21.7989C29.1108 21.7989 30.1411 22.9716 30.1191 24.4019C30.1191 25.8301 29.1108 27.0029 27.8325 27.0029Z" fill="white"/>
-              </svg>
-            </button>
-          </div>
-
-          <div className="flex justify-center">
-            <Link href='/' className="next"><div>Start Saving</div>
-              <div>
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="24"
-                  height="24"
-                  viewBox="0 0 24 24"
-                  fill="none"
-                >
-                  <path
-                    d="M5.20874 13H16.3787L11.4987 17.88C11.1087 18.27 11.1087 18.91 11.4987 19.3C11.8887 19.69 12.5187 19.69 12.9087 19.3L19.4987 12.71C19.8887 12.32 19.8887 11.69 19.4987 11.3L12.9187 4.69996C12.7319 4.5127 12.4783 4.40747 12.2137 4.40747C11.9492 4.40747 11.6956 4.5127 11.5087 4.69996C11.1187 5.08996 11.1187 5.71996 11.5087 6.10996L16.3787 11H5.20874C4.65874 11 4.20874 11.45 4.20874 12C4.20874 12.55 4.65874 13 5.20874 13Z"
-                    fill="#8E94E9"
-                  />
-                </svg>
-              </div>
-            </Link>
-          </div>
-        </div>
-      </div>
-      <IllustrationSix />
-    </div>
-  )
-}
