+++ fwb/app/layout.tsx
@@ -1,29 +1,11 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
 import "./globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
-import { dark } from "@clerk/themes";
-
-const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: "Friends With Benefits",
   description: "Discounts Have Never Been Easier To Find!",
 };
-
-// export default function RootLayout({
-//   children,
-// }: {
-//   children: React.ReactNode
-// }) {
-//   return (
-//     <ClerkProvider>
-//       <html lang="en">
-//         <body className={inter.className}>{children}</body>
-//       </html>
-//     </ClerkProvider>
-//   )
-// }
 
 export default function RootLayout({
   children,
@@ -31,18 +13,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <ClerkProvider
-      appearance={{
-        signIn: {
-          variables: {
-            colorBackground: "#bbbef2",
-            colorText: "white",
-
-            // baseTheme: dark // dark 테마를 적용합니다.
-          },
-        },
-      }}
-    >
+    <ClerkProvider>
       <html lang="en">
         <body>{children}</body>
       </html>
