--- fwb/components/ui/fre/Profilepicture.tsx
@@ -1,204 +1,0 @@
-import React from 'react'
-
-export default function Profilepicture() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="234" height="322" viewBox="0 0 234 322" fill="none">
-<g clip-path="url(#clip0_1631_30438)">
-<rect x="38" width="160" height="160" rx="80" fill="white"/>
-<path d="M191.599 112.791L191.595 112.801C190.646 114.92 189.622 117.004 188.488 119.049C187.915 120.107 187.314 121.138 186.703 122.138C186.703 122.139 186.703 122.139 186.702 122.14L183.81 120.375C182.76 122.097 181.644 123.755 180.479 125.363L191.599 112.791ZM191.599 112.791L191.603 112.782M191.599 112.791L191.603 112.782M191.603 112.782C195.957 102.91 198.39 91.9766 198.39 80.5M191.603 112.782L198.39 80.5M198.39 80.5C198.39 36.3746 162.626 0.611295 118.5 0.611295C74.3748 0.611295 38.6113 36.3746 38.6113 80.5C38.6113 92.401 41.2145 103.731 45.9296 113.917M198.39 80.5L45.9296 113.917M45.9296 113.917C48.0572 118.549 50.6274 122.934 53.5525 127.022L45.9296 113.917ZM182.472 128.375C182.543 128.29 182.626 128.184 182.71 128.06L53.553 127.023C53.8229 127.401 54.0876 127.754 54.3537 128.103C55.5782 129.757 56.8754 131.343 58.183 132.868L58.1899 132.876L58.1968 132.884C61.1221 136.255 64.3238 139.372 67.7691 142.185C70.2058 144.226 72.763 146.096 75.4401 147.795C82.8088 152.524 91.0166 156.07 99.772 158.178L99.7742 158.178C105.773 159.618 112.06 160.389 118.5 160.389H118.54C119.002 160.389 119.57 160.389 120.157 160.354C123.452 160.3 126.686 160.044 129.874 159.584C139.027 158.281 147.649 155.435 155.515 151.322L155.525 151.317L155.535 151.312C156.621 150.735 157.691 150.141 158.745 149.529C160.303 148.628 161.814 147.666 163.273 146.682L163.274 146.681C165.29 145.32 167.226 143.863 169.084 142.346L169.095 142.337C174.081 138.232 178.586 133.551 182.472 128.375Z" fill="#85E0D3" stroke="white" stroke-width="6.77741"/>
-<path d="M166.944 139.718C165.155 141.178 163.301 142.573 161.381 143.869C159.97 144.821 158.526 145.74 157.049 146.593C148.713 151.467 139.376 154.798 129.399 156.225C126.33 156.669 123.213 156.915 120.029 156.964C119.52 156.997 119.012 156.997 118.503 156.997C112.333 156.997 106.311 156.258 100.568 154.88C92.1825 152.862 84.3224 149.465 77.2663 144.936C74.7065 143.311 72.2615 141.523 69.9313 139.57C66.6331 136.879 63.5645 133.892 60.7585 130.66C59.4786 129.167 58.2315 127.641 57.0664 126.065C62.2846 113.709 71.8184 96.0034 84.8147 91.6714C105.277 84.8451 148.877 94.461 148.877 94.461C148.877 94.461 163.809 101.714 166.944 139.718Z" fill="#FC7A82"/>
-<path d="M157.048 146.593C148.712 151.467 139.375 154.798 129.398 156.226C126.33 156.669 123.212 156.915 120.028 156.964C119.52 156.997 119.011 156.997 118.502 156.997C112.332 156.997 106.31 156.258 100.567 154.88C92.1818 152.862 84.3217 149.465 77.2657 144.936C74.7058 143.311 72.2608 141.523 69.9307 139.57C70.0619 119.567 71.5716 101.287 76.4124 96.9224C78.841 94.7236 81.7454 93.0006 84.814 91.6715C91.9192 88.6029 99.8614 87.5691 105.08 87.2245C108.542 86.9948 110.79 87.0769 110.79 87.0769C110.79 87.0769 115.434 87.2574 121.669 87.9466C123.392 88.1599 125.132 88.3732 126.953 88.6357C137.57 90.1618 150.008 93.1319 153.225 98.7438C155.998 103.568 156.884 125.179 157.048 146.593Z" fill="#FC7A82"/>
-<path d="M133.191 47.0087L101.57 70.9499C101.57 70.9499 84.34 53.6545 91.3468 43.7761C95.6953 37.6226 104.491 34.669 110.513 33.307C111.202 33.1593 111.842 33.028 112.449 32.9132C115.042 32.4045 116.781 32.2568 116.781 32.2568L116.798 32.3553C117.897 31.945 128.859 24.0194 134.865 26.6448C143.117 30.2385 133.191 47.0087 133.191 47.0087Z" fill="black"/>
-<path d="M104.85 74.6895L103.808 91.6796C103.808 91.6796 106.573 96.7517 116.461 98.1695C123.829 99.2262 126.531 97.7117 126.531 97.7117L126.028 89.0443L125.525 80.5902L107.821 75.5411L104.85 74.6895Z" fill="#FCC898"/>
-<path d="M107.818 75.5413C107.833 76.0073 107.976 76.5668 108.212 77.2117C111.492 85.9316 121.775 88.3996 126.026 89.0445L125.524 80.5904L107.818 75.5413Z" fill="black"/>
-<path d="M131.686 42.0974C131.686 42.0974 139.295 63.9463 135.055 75.5657C130.813 87.1851 120.139 84.4841 113.841 81.6896C107.545 78.8968 104.845 74.6944 104.845 74.6944C104.845 74.6944 98.6407 76.2795 95.029 70.3443C91.4173 64.4107 94.4038 58.144 97.9712 56.9297C101.539 55.7154 105.452 58.5247 105.452 58.5247L102.707 52.1333C102.704 52.1349 118.645 21.9566 131.686 42.0974Z" fill="#FCC898"/>
-<path d="M104.304 67.8025C104.121 67.8025 103.936 67.7417 103.785 67.617L96.4681 61.6424C96.1169 61.3569 96.0661 60.8384 96.3516 60.4872C96.6421 60.136 97.1573 60.0868 97.5068 60.3707L104.824 66.3453C105.175 66.6308 105.226 67.1494 104.94 67.5005C104.776 67.6991 104.542 67.8025 104.304 67.8025Z" fill="black"/>
-<path d="M97.3306 67.2793C96.9467 67.2793 96.6037 67.007 96.5266 66.6148C96.4413 66.1701 96.7317 65.7401 97.178 65.6532L102.298 64.6637C102.738 64.5767 103.171 64.8672 103.259 65.3135C103.345 65.7582 103.054 66.1881 102.61 66.2735L97.4898 67.2629C97.4357 67.2761 97.3831 67.2793 97.3306 67.2793Z" fill="black"/>
-<path d="M116.974 51.239C116.782 51.239 116.59 51.1717 116.434 51.0355C116.092 50.7368 116.058 50.22 116.357 49.8786C118.903 46.9644 122.028 47.7225 123.324 48.5282C123.708 48.7677 123.826 49.2748 123.586 49.6588C123.344 50.0427 122.842 50.1592 122.458 49.9229C122.188 49.7605 119.806 48.4281 117.589 50.9584C117.428 51.1438 117.202 51.239 116.974 51.239Z" fill="black"/>
-<path d="M133.478 50.5417C133.277 50.5417 133.076 50.4679 132.916 50.3202C131.3 48.8056 129.431 49.6967 129.354 49.7344C128.949 49.9362 128.458 49.7705 128.254 49.3685C128.053 48.9648 128.212 48.4758 128.612 48.2707C128.73 48.21 131.558 46.807 134.039 49.1207C134.369 49.4308 134.387 49.9494 134.076 50.2808C133.916 50.4548 133.697 50.5417 133.478 50.5417Z" fill="black"/>
-<path d="M129.625 66.5273C129.531 66.5273 129.436 66.5109 129.343 66.4781C128.916 66.3222 128.698 65.8513 128.854 65.4246C131.013 59.5288 128.204 56.1124 128.082 55.9696C127.79 55.625 127.831 55.1081 128.174 54.8144C128.517 54.519 129.031 54.5568 129.326 54.8997C129.474 55.0704 132.899 59.1465 130.395 65.9875C130.273 66.3222 129.96 66.5273 129.625 66.5273Z" fill="black"/>
-<path d="M132.9 58.7003C132.519 58.7003 132.18 58.4344 132.099 58.0472L131.274 54.0843C131.261 54.0286 131.256 53.9711 131.256 53.9153C131.256 53.5346 131.522 53.1933 131.909 53.1129C132.342 53.021 132.787 53.3033 132.879 53.748L133.704 57.7108C133.717 57.7666 133.722 57.824 133.722 57.8782C133.722 58.2589 133.456 58.6002 133.069 58.6806C133.012 58.6937 132.956 58.7003 132.9 58.7003Z" fill="black"/>
-<path d="M120.487 58.9871C120.108 58.9871 119.766 58.7212 119.686 58.334L118.861 54.3728C118.767 53.9281 119.053 53.4949 119.497 53.403C119.932 53.3095 120.375 53.595 120.467 54.038L121.292 57.9992C121.386 58.4439 121.1 58.8771 120.656 58.969C120.6 58.9822 120.544 58.9871 120.487 58.9871Z" fill="black"/>
-<path d="M133.183 47.007C133.183 47.007 131.547 43.479 127.312 43.1902C123.075 42.9014 120.419 45.7336 120.419 45.7336C120.419 45.7336 121.84 52.9849 115.844 56.4226C109.848 59.8603 105.453 58.5263 105.453 58.5263L101.364 56.7803L98.6729 50.4841L111.566 38.1935L127.724 30.8931L135.256 38.6235L133.183 47.007Z" fill="black"/>
-<path d="M124.936 73.0355C124.275 73.0355 120.362 72.8682 118.835 69.0186C118.668 68.5968 118.875 68.121 119.295 67.9536C119.715 67.7862 120.192 67.993 120.36 68.4147C121.571 71.4717 124.905 71.393 124.971 71.3946C125.439 71.3946 125.803 71.7425 125.819 72.1855C125.834 72.6384 125.48 73.0175 125.027 73.0339C125.015 73.0339 124.986 73.0355 124.936 73.0355Z" fill="black"/>
-<path d="M121.892 99.4887C110.768 99.4887 102.303 91.7845 101.893 91.4055C101.561 91.097 101.54 90.5801 101.848 90.247C102.158 89.9139 102.674 89.8942 103.008 90.201C103.117 90.3028 114.198 100.376 127.101 97.2226C127.537 97.1143 127.984 97.385 128.092 97.8248C128.199 98.2645 127.929 98.7092 127.49 98.8159C125.57 99.2852 123.696 99.4887 121.892 99.4887Z" fill="white"/>
-<path d="M180.482 125.363C180.269 125.659 180.072 125.937 179.859 126.216C176.117 131.221 171.769 135.75 166.944 139.721C165.172 141.165 163.334 142.56 161.431 143.84L147.188 137.769L152.289 97.4559L180.482 125.363Z" fill="#FC7A82"/>
-</g>
-<rect x="11.5" y="168" width="213" height="45" rx="22.5" fill="#8E94E9"/>
-<path d="M41.6566 184H44.4666L49.5566 198H46.9666L46.0666 195.52H40.0666L39.1666 198H36.5766L41.6566 184ZM40.9566 193.08H45.1666L43.0566 187.31L40.9566 193.08ZM58.5108 183H60.9508V198H58.5108V196.75C58.1908 197.197 57.7875 197.56 57.3008 197.84C56.8141 198.12 56.2508 198.26 55.6108 198.26C54.8841 198.26 54.2041 198.123 53.5708 197.85C52.9375 197.577 52.3775 197.2 51.8908 196.72C51.4108 196.233 51.0341 195.673 50.7608 195.04C50.4875 194.407 50.3508 193.727 50.3508 193C50.3508 192.273 50.4875 191.593 50.7608 190.96C51.0341 190.327 51.4108 189.77 51.8908 189.29C52.3775 188.803 52.9375 188.423 53.5708 188.15C54.2041 187.877 54.8841 187.74 55.6108 187.74C56.2508 187.74 56.8141 187.88 57.3008 188.16C57.7875 188.433 58.1908 188.797 58.5108 189.25V183ZM55.6408 195.91C56.1608 195.91 56.6208 195.78 57.0208 195.52C57.4208 195.26 57.7341 194.91 57.9608 194.47C58.1875 194.023 58.3008 193.533 58.3008 193C58.3008 192.46 58.1875 191.97 57.9608 191.53C57.7341 191.083 57.4175 190.73 57.0108 190.47C56.6108 190.21 56.1541 190.08 55.6408 190.08C55.1275 190.08 54.6575 190.213 54.2308 190.48C53.8108 190.74 53.4741 191.09 53.2208 191.53C52.9741 191.97 52.8508 192.46 52.8508 193C52.8508 193.54 52.9775 194.03 53.2308 194.47C53.4841 194.91 53.8208 195.26 54.2408 195.52C54.6675 195.78 55.1341 195.91 55.6408 195.91ZM71.1569 183H73.5969V198H71.1569V196.75C70.8369 197.197 70.4335 197.56 69.9469 197.84C69.4602 198.12 68.8969 198.26 68.2569 198.26C67.5302 198.26 66.8502 198.123 66.2169 197.85C65.5835 197.577 65.0235 197.2 64.5369 196.72C64.0569 196.233 63.6802 195.673 63.4069 195.04C63.1335 194.407 62.9969 193.727 62.9969 193C62.9969 192.273 63.1335 191.593 63.4069 190.96C63.6802 190.327 64.0569 189.77 64.5369 189.29C65.0235 188.803 65.5835 188.423 66.2169 188.15C66.8502 187.877 67.5302 187.74 68.2569 187.74C68.8969 187.74 69.4602 187.88 69.9469 188.16C70.4335 188.433 70.8369 188.797 71.1569 189.25V183ZM68.2869 195.91C68.8069 195.91 69.2669 195.78 69.6669 195.52C70.0669 195.26 70.3802 194.91 70.6069 194.47C70.8335 194.023 70.9469 193.533 70.9469 193C70.9469 192.46 70.8335 191.97 70.6069 191.53C70.3802 191.083 70.0635 190.73 69.6569 190.47C69.2569 190.21 68.8002 190.08 68.2869 190.08C67.7735 190.08 67.3035 190.213 66.8769 190.48C66.4569 190.74 66.1202 191.09 65.8669 191.53C65.6202 191.97 65.4969 192.46 65.4969 193C65.4969 193.54 65.6235 194.03 65.8769 194.47C66.1302 194.91 66.4669 195.26 66.8869 195.52C67.3135 195.78 67.7802 195.91 68.2869 195.91ZM84.0439 203H81.6039V188H84.0439V189.27C84.3706 188.817 84.7706 188.45 85.2439 188.17C85.7239 187.883 86.2839 187.74 86.9239 187.74C87.6572 187.74 88.3406 187.877 88.9739 188.15C89.6072 188.423 90.1639 188.803 90.6439 189.29C91.1306 189.77 91.5072 190.327 91.7739 190.96C92.0472 191.593 92.1839 192.273 92.1839 193C92.1839 193.727 92.0472 194.41 91.7739 195.05C91.5072 195.69 91.1306 196.253 90.6439 196.74C90.1639 197.22 89.6072 197.597 88.9739 197.87C88.3406 198.143 87.6572 198.28 86.9239 198.28C86.2839 198.28 85.7239 198.14 85.2439 197.86C84.7706 197.573 84.3706 197.203 84.0439 196.75V203ZM86.8939 190.09C86.3872 190.09 85.9339 190.223 85.5339 190.49C85.1339 190.75 84.8206 191.1 84.5939 191.54C84.3672 191.98 84.2539 192.467 84.2539 193C84.2539 193.533 84.3672 194.023 84.5939 194.47C84.8206 194.91 85.1339 195.263 85.5339 195.53C85.9339 195.79 86.3872 195.92 86.8939 195.92C87.4072 195.92 87.8772 195.79 88.3039 195.53C88.7306 195.27 89.0706 194.92 89.3239 194.48C89.5772 194.033 89.7039 193.54 89.7039 193C89.7039 192.467 89.5772 191.98 89.3239 191.54C89.0706 191.1 88.7306 190.75 88.3039 190.49C87.8839 190.223 87.4139 190.09 86.8939 190.09ZM94.1623 198L94.1523 188H96.5923L96.6023 188.89C96.9423 188.53 97.349 188.247 97.8223 188.04C98.2957 187.827 98.8057 187.72 99.3523 187.72C99.719 187.72 100.086 187.773 100.452 187.88L99.4923 190.34C99.239 190.24 98.9857 190.19 98.7323 190.19C98.339 190.19 97.979 190.287 97.6523 190.48C97.3323 190.667 97.0757 190.923 96.8823 191.25C96.6957 191.57 96.6023 191.927 96.6023 192.32V198H94.1623ZM105.905 198.26C104.985 198.26 104.145 198.023 103.385 197.55C102.632 197.077 102.029 196.443 101.575 195.65C101.129 194.85 100.905 193.967 100.905 193C100.905 192.267 101.035 191.583 101.295 190.95C101.555 190.31 101.912 189.75 102.365 189.27C102.825 188.783 103.359 188.403 103.965 188.13C104.572 187.857 105.219 187.72 105.905 187.72C106.825 187.72 107.662 187.957 108.415 188.43C109.175 188.903 109.779 189.54 110.225 190.34C110.679 191.14 110.905 192.027 110.905 193C110.905 193.727 110.775 194.407 110.515 195.04C110.255 195.673 109.895 196.233 109.435 196.72C108.982 197.2 108.452 197.577 107.845 197.85C107.245 198.123 106.599 198.26 105.905 198.26ZM105.905 195.82C106.392 195.82 106.829 195.69 107.215 195.43C107.602 195.17 107.905 194.827 108.125 194.4C108.352 193.973 108.465 193.507 108.465 193C108.465 192.48 108.349 192.007 108.115 191.58C107.889 191.147 107.579 190.803 107.185 190.55C106.799 190.29 106.372 190.16 105.905 190.16C105.425 190.16 104.992 190.29 104.605 190.55C104.219 190.81 103.912 191.157 103.685 191.59C103.459 192.023 103.345 192.493 103.345 193C103.345 193.527 103.462 194.003 103.695 194.43C103.929 194.857 104.239 195.197 104.625 195.45C105.019 195.697 105.445 195.82 105.905 195.82ZM115.927 185.9V188H118.657V190.44H115.927V198H113.487V190.44H112.397V188H113.487V185.9C113.487 185.327 113.627 184.803 113.907 184.33C114.194 183.85 114.577 183.47 115.057 183.19C115.537 182.903 116.067 182.76 116.647 182.76C117.067 182.76 117.484 182.847 117.897 183.02C118.31 183.193 118.677 183.46 118.997 183.82L117.237 185.54C117.177 185.427 117.09 185.343 116.977 185.29C116.87 185.23 116.76 185.2 116.647 185.2C116.454 185.2 116.284 185.27 116.137 185.41C115.997 185.543 115.927 185.707 115.927 185.9ZM120.626 188H123.066V198H120.626V188ZM121.866 186.71C121.506 186.71 121.202 186.593 120.956 186.36C120.709 186.12 120.586 185.82 120.586 185.46C120.586 185.107 120.709 184.81 120.956 184.57C121.202 184.33 121.502 184.21 121.856 184.21C122.209 184.21 122.506 184.33 122.746 184.57C122.992 184.81 123.116 185.107 123.116 185.46C123.116 185.82 122.996 186.12 122.756 186.36C122.516 186.593 122.219 186.71 121.866 186.71ZM125.909 198V183H128.349V198H125.909ZM135.396 198.26C134.476 198.26 133.636 198.023 132.876 197.55C132.123 197.077 131.519 196.44 131.066 195.64C130.619 194.84 130.396 193.957 130.396 192.99C130.396 192.263 130.526 191.583 130.786 190.95C131.046 190.31 131.403 189.75 131.856 189.27C132.316 188.783 132.849 188.403 133.456 188.13C134.063 187.857 134.709 187.72 135.396 187.72C136.176 187.72 136.889 187.887 137.536 188.22C138.189 188.547 138.743 188.997 139.196 189.57C139.649 190.143 139.979 190.797 140.186 191.53C140.393 192.263 140.439 193.03 140.326 193.83H133.016C133.109 194.203 133.263 194.54 133.476 194.84C133.689 195.133 133.959 195.37 134.286 195.55C134.613 195.723 134.983 195.813 135.396 195.82C135.823 195.827 136.209 195.727 136.556 195.52C136.909 195.307 137.203 195.02 137.436 194.66L139.926 195.24C139.519 196.127 138.913 196.853 138.106 197.42C137.299 197.98 136.396 198.26 135.396 198.26ZM132.936 192H137.856C137.783 191.6 137.626 191.24 137.386 190.92C137.153 190.593 136.863 190.333 136.516 190.14C136.169 189.947 135.796 189.85 135.396 189.85C134.996 189.85 134.626 189.947 134.286 190.14C133.946 190.327 133.656 190.583 133.416 190.91C133.183 191.23 133.023 191.593 132.936 192ZM150.232 203H147.792V188H150.232V189.27C150.559 188.817 150.959 188.45 151.432 188.17C151.912 187.883 152.472 187.74 153.112 187.74C153.846 187.74 154.529 187.877 155.162 188.15C155.796 188.423 156.352 188.803 156.832 189.29C157.319 189.77 157.696 190.327 157.962 190.96C158.236 191.593 158.372 192.273 158.372 193C158.372 193.727 158.236 194.41 157.962 195.05C157.696 195.69 157.319 196.253 156.832 196.74C156.352 197.22 155.796 197.597 155.162 197.87C154.529 198.143 153.846 198.28 153.112 198.28C152.472 198.28 151.912 198.14 151.432 197.86C150.959 197.573 150.559 197.203 150.232 196.75V203ZM153.082 190.09C152.576 190.09 152.122 190.223 151.722 190.49C151.322 190.75 151.009 191.1 150.782 191.54C150.556 191.98 150.442 192.467 150.442 193C150.442 193.533 150.556 194.023 150.782 194.47C151.009 194.91 151.322 195.263 151.722 195.53C152.122 195.79 152.576 195.92 153.082 195.92C153.596 195.92 154.066 195.79 154.492 195.53C154.919 195.27 155.259 194.92 155.512 194.48C155.766 194.033 155.892 193.54 155.892 193C155.892 192.467 155.766 191.98 155.512 191.54C155.259 191.1 154.919 190.75 154.492 190.49C154.072 190.223 153.602 190.09 153.082 190.09ZM169.344 191.68V198H166.904V192.32C166.904 191.927 166.808 191.57 166.614 191.25C166.421 190.923 166.161 190.667 165.834 190.48C165.514 190.287 165.158 190.19 164.764 190.19C164.371 190.19 164.011 190.287 163.684 190.48C163.364 190.667 163.108 190.923 162.914 191.25C162.728 191.57 162.634 191.927 162.634 192.32V198H160.194L160.184 183H162.624L162.634 188.89C162.974 188.53 163.381 188.247 163.854 188.04C164.328 187.827 164.838 187.72 165.384 187.72C166.111 187.72 166.774 187.9 167.374 188.26C167.974 188.613 168.451 189.09 168.804 189.69C169.164 190.283 169.344 190.947 169.344 191.68ZM175.941 198.26C175.021 198.26 174.181 198.023 173.421 197.55C172.668 197.077 172.065 196.443 171.611 195.65C171.165 194.85 170.941 193.967 170.941 193C170.941 192.267 171.071 191.583 171.331 190.95C171.591 190.31 171.948 189.75 172.401 189.27C172.861 188.783 173.395 188.403 174.001 188.13C174.608 187.857 175.255 187.72 175.941 187.72C176.861 187.72 177.698 187.957 178.451 188.43C179.211 188.903 179.815 189.54 180.261 190.34C180.715 191.14 180.941 192.027 180.941 193C180.941 193.727 180.811 194.407 180.551 195.04C180.291 195.673 179.931 196.233 179.471 196.72C179.018 197.2 178.488 197.577 177.881 197.85C177.281 198.123 176.635 198.26 175.941 198.26ZM175.941 195.82C176.428 195.82 176.865 195.69 177.251 195.43C177.638 195.17 177.941 194.827 178.161 194.4C178.388 193.973 178.501 193.507 178.501 193C178.501 192.48 178.385 192.007 178.151 191.58C177.925 191.147 177.615 190.803 177.221 190.55C176.835 190.29 176.408 190.16 175.941 190.16C175.461 190.16 175.028 190.29 174.641 190.55C174.255 190.81 173.948 191.157 173.721 191.59C173.495 192.023 173.381 192.493 173.381 193C173.381 193.527 173.498 194.003 173.731 194.43C173.965 194.857 174.275 195.197 174.661 195.45C175.055 195.697 175.481 195.82 175.941 195.82ZM188.258 190.44H186.318V198H183.878V190.44H182.398V188H183.878V184.86H186.318V188H188.258V190.44ZM194.632 198.26C193.712 198.26 192.872 198.023 192.112 197.55C191.359 197.077 190.755 196.443 190.302 195.65C189.855 194.85 189.632 193.967 189.632 193C189.632 192.267 189.762 191.583 190.022 190.95C190.282 190.31 190.639 189.75 191.092 189.27C191.552 188.783 192.085 188.403 192.692 188.13C193.299 187.857 193.945 187.72 194.632 187.72C195.552 187.72 196.389 187.957 197.142 188.43C197.902 188.903 198.505 189.54 198.952 190.34C199.405 191.14 199.632 192.027 199.632 193C199.632 193.727 199.502 194.407 199.242 195.04C198.982 195.673 198.622 196.233 198.162 196.72C197.709 197.2 197.179 197.577 196.572 197.85C195.972 198.123 195.325 198.26 194.632 198.26ZM194.632 195.82C195.119 195.82 195.555 195.69 195.942 195.43C196.329 195.17 196.632 194.827 196.852 194.4C197.079 193.973 197.192 193.507 197.192 193C197.192 192.48 197.075 192.007 196.842 191.58C196.615 191.147 196.305 190.803 195.912 190.55C195.525 190.29 195.099 190.16 194.632 190.16C194.152 190.16 193.719 190.29 193.332 190.55C192.945 190.81 192.639 191.157 192.412 191.59C192.185 192.023 192.072 192.493 192.072 193C192.072 193.527 192.189 194.003 192.422 194.43C192.655 194.857 192.965 195.197 193.352 195.45C193.745 195.697 194.172 195.82 194.632 195.82Z" fill="white"/>
-<rect x="165.5" y="15.5" width="37" height="37" rx="18.5" fill="white"/>
-<rect x="165.5" y="15.5" width="37" height="37" rx="18.5" stroke="#8E94E9" stroke-width="3"/>
-<path d="M189.596 25.9998C188.959 25.9998 188.348 26.253 187.897 26.7037L177.133 37.4676L176 41.9998L180.532 40.8667L191.296 30.1029C191.747 29.6521 192 29.0408 192 28.4033C192 27.7658 191.747 27.1545 191.296 26.7037C190.845 26.253 190.234 25.9998 189.596 25.9998Z" fill="#8E94E9" stroke="#8E94E9" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M43.1289 258.121L43.1229 258.128L43.1227 258.128L43.1167 258.136L43.1049 258.151C43.0942 258.165 43.0734 258.192 43.045 258.224C41.2455 260.495 39.0342 262.432 36.5139 263.899C34.7801 264.925 32.9037 265.734 30.9201 266.286M43.1289 258.121L30.6522 265.322M43.1289 258.121L43.1379 258.11M43.1289 258.121L43.1379 258.11M30.9201 266.286L30.6522 265.322M30.9201 266.286C30.9205 266.285 30.9208 266.285 30.9212 266.285L30.6522 265.322M30.9201 266.286C29.5812 266.659 28.1959 266.917 26.7697 267.044M30.6522 265.322C29.369 265.68 28.0422 265.927 26.6767 266.048M26.7697 267.044C26.7714 267.044 26.7731 267.044 26.7748 267.044L26.6767 266.048M26.7697 267.044C26.0502 267.115 25.317 267.15 24.5799 267.15C23.7486 267.15 22.9241 267.11 22.107 267.024M26.7697 267.044C26.7681 267.044 26.7666 267.044 26.765 267.045L26.6767 266.048M26.6767 266.048L22.107 267.024M22.107 267.024C22.1073 267.024 22.1075 267.024 22.1078 267.024L22.212 266.029L22.107 267.024ZM43.1379 258.11C44.4677 256.416 45.5701 254.535 46.3996 252.517C47.5333 249.758 48.155 246.738 48.155 243.575M43.1379 258.11L48.155 243.575M15.7808 265.448C14.6382 264.991 13.5372 264.447 12.4887 263.812C12.1848 263.633 11.8911 263.449 11.5985 263.256C11.5665 263.236 11.5379 263.215 11.5131 263.196C11.2675 263.035 11.0147 262.863 10.7682 262.684L10.7602 262.678L10.7602 262.678C10.727 262.653 10.6866 262.624 10.6427 262.592C10.5535 262.527 10.4496 262.451 10.3615 262.384C9.57616 261.789 8.82983 261.149 8.12305 260.458C8.01244 260.352 7.89514 260.235 7.78317 260.113C6.2419 258.554 4.92661 256.789 3.87276 254.858L15.7808 265.448ZM15.7808 265.448C15.7925 265.453 15.8029 265.457 15.8117 265.461C15.8142 265.462 15.8166 265.463 15.819 265.464L15.7808 265.448ZM48.155 243.575C48.155 230.554 37.6013 220 24.5799 220C11.5588 220 1 230.553 1 243.575C1 247.66 2.04175 251.504 3.87265 254.857L48.155 243.575ZM18.8315 266.439C18.8336 266.439 18.8356 266.44 18.8377 266.44L19.0693 265.467L18.8247 266.437C18.827 266.438 18.8292 266.438 18.8315 266.439ZM18.4949 266.354C18.498 266.355 18.5011 266.356 18.5041 266.357L18.7594 265.39L18.4847 266.351C18.4881 266.352 18.4915 266.353 18.4949 266.354Z" fill="#F6FF82" stroke="white" stroke-width="2"/>
-<path d="M29.443 231.219C29.443 231.219 31.8666 230.824 31.916 237.236C31.9654 243.649 31.9161 245.938 31.9161 245.938H25.9658L27.0945 232.97L29.443 231.219Z" fill="black"/>
-<path d="M19.9201 229.342C19.9201 229.342 17.5812 232.343 16.5028 236.596C15.6399 239.998 13.8643 251.472 13.8643 251.472L20.401 252.233L21.7849 232.916L19.9201 229.342Z" fill="black"/>
-<path d="M22.2843 249.315C20.8486 249.315 19.852 248.905 19.8331 248.896C19.7208 248.848 19.668 248.718 19.716 248.605C19.7639 248.492 19.8956 248.44 20.007 248.488C20.0404 248.502 23.4417 249.893 26.6623 247.296C26.7572 247.219 26.8977 247.234 26.9746 247.329C27.0516 247.425 27.0366 247.564 26.9417 247.642C25.2963 248.968 23.6048 249.315 22.2843 249.315Z" fill="#F7C537"/>
-<path d="M36.0084 263.037C34.3474 264.02 32.5509 264.794 30.6527 265.322C29.3695 265.68 28.0427 265.927 26.6771 266.048C25.9895 266.116 25.2874 266.15 24.5804 266.15C23.7814 266.15 22.9921 266.111 22.2125 266.029C21.704 265.976 21.2004 265.903 20.7017 265.816C20.1496 265.719 19.6073 265.603 19.0698 265.467C18.9681 265.443 18.8616 265.419 18.7599 265.39C17.8979 265.162 17.0505 264.881 16.2322 264.552C16.2128 264.542 16.1934 264.538 16.1741 264.528C15.07 264.087 14.0095 263.564 13.0023 262.954C12.707 262.78 12.4212 262.601 12.1355 262.412C12.1258 262.407 12.121 262.402 12.1113 262.392C12.3341 259.569 12.8667 256.262 14.058 253.502C14.087 253.434 14.1161 253.366 14.15 253.299C15.7818 249.657 18.3483 247.928 20.1158 247.134C21.2682 246.621 22.0818 246.495 22.0818 246.495C22.0818 246.495 24.0913 246.117 26.527 246.083C29.6261 246.04 33.4128 246.548 34.6137 249.115C35.5338 251.086 35.8921 256.969 36.0084 263.037Z" fill="white"/>
-<path d="M20.5559 246.558C20.4155 246.65 20.2703 246.844 20.1153 247.134C18.9774 249.275 17.5343 256.548 16.1736 264.528C15.0696 264.087 14.0091 263.564 13.0019 262.954C12.7065 262.78 12.4208 262.601 12.1351 262.412C12.1254 262.407 12.1206 262.402 12.1109 262.392C11.8591 262.228 11.6024 262.054 11.3555 261.874C11.462 260.795 11.5201 260.17 11.5201 260.17L14.4158 251.066C14.4158 251.066 15.7329 249.27 17.3358 248.127C18.9337 246.989 20.5559 246.558 20.5559 246.558Z" fill="#2F2F2F"/>
-<path d="M42.284 257.574C40.5553 259.763 38.4295 261.627 36.0083 263.036C34.3473 264.019 32.5508 264.794 30.6526 265.322C29.3694 265.68 28.0426 265.927 26.6771 266.048C26.8611 255.942 26.9628 246.786 26.6093 246.461C26.5076 246.369 26.4979 246.234 26.527 246.083C26.619 245.706 27.0112 245.251 27.0112 245.251C35.9308 244.897 39.5868 251.086 40.1727 252.441C41.0056 254.369 41.6932 255.938 42.284 257.574Z" fill="#2F2F2F"/>
-<path d="M26.9133 241.522L26.8009 246.581C26.8009 246.581 25.8561 248.015 22.8917 248.19C20.6831 248.32 19.9199 247.804 19.9199 247.804L20.2841 245.249L20.6424 242.757L26.0121 241.701L26.9133 241.522Z" fill="#FBD7C6"/>
-<path d="M26.0117 241.701C25.9957 241.839 25.9395 242 25.8538 242.185C24.666 244.687 21.5592 245.163 20.2832 245.249L20.6415 242.757L26.0117 241.701Z" fill="black"/>
-<path d="M19.7715 231.202C19.7715 231.202 16.976 237.485 17.944 241.032C18.9124 244.579 22.1418 244.044 24.0763 243.372C26.0109 242.7 26.9149 241.523 26.9149 241.523C26.9149 241.523 28.7139 242.146 29.9308 240.478C31.1476 238.81 30.4184 236.879 29.3918 236.431C28.3652 235.983 27.1362 236.718 27.1362 236.718L28.1081 234.893C28.1081 234.893 24.134 225.559 19.7715 231.202Z" fill="#FBD7C6"/>
-<path d="M20.0082 239.252C19.9046 239.252 19.8092 239.185 19.7772 239.082C19.3545 237.715 20.3147 236.328 20.3558 236.27C20.4328 236.16 20.5835 236.135 20.6929 236.211C20.8023 236.288 20.8284 236.439 20.7519 236.548C20.7432 236.56 19.8876 237.801 20.2392 238.938C20.2789 239.066 20.2072 239.202 20.0793 239.241C20.0561 239.249 20.0319 239.252 20.0082 239.252Z" fill="black"/>
-<path d="M27.4052 239.716C27.3456 239.716 27.2861 239.694 27.2391 239.65C27.1418 239.558 27.1374 239.405 27.2289 239.308L28.9407 237.494C29.0317 237.397 29.1852 237.392 29.2826 237.483C29.3799 237.575 29.3843 237.728 29.2922 237.826L27.5805 239.64C27.5335 239.69 27.4696 239.716 27.4052 239.716Z" fill="black"/>
-<path d="M29.2175 239.23C29.2054 239.23 29.1928 239.229 29.1807 239.227L27.9967 239.048C27.865 239.028 27.7735 238.904 27.7938 238.772C27.8137 238.64 27.9381 238.547 28.0694 238.569L29.2533 238.748C29.385 238.768 29.4766 238.892 29.4562 239.024C29.4383 239.143 29.3352 239.23 29.2175 239.23Z" fill="black"/>
-<path d="M23.9586 234.651C23.8709 234.651 23.7867 234.603 23.7436 234.52C23.4705 233.988 22.7141 234.145 22.7064 234.146C22.5751 234.177 22.4463 234.091 22.4187 233.961C22.3902 233.83 22.473 233.701 22.6042 233.673C22.6502 233.664 23.732 233.44 24.1736 234.299C24.2351 234.418 24.1881 234.563 24.069 234.625C24.0337 234.642 23.9959 234.651 23.9586 234.651Z" fill="black"/>
-<path d="M18.7705 234.123C18.7202 234.123 18.6688 234.107 18.6257 234.074C18.5192 233.995 18.4964 233.844 18.5754 233.738C18.8146 233.414 19.4974 232.951 20.3646 233.516C20.477 233.589 20.5084 233.739 20.4353 233.851C20.3627 233.964 20.2121 233.995 20.1007 233.921C19.4131 233.475 19.0073 233.969 18.9633 234.027C18.9158 234.09 18.8436 234.123 18.7705 234.123Z" fill="black"/>
-<path d="M22.7494 248.523C21.1558 248.523 19.8576 248.04 19.8333 248.031C19.7084 247.983 19.6459 247.843 19.6939 247.718C19.7418 247.593 19.8818 247.532 20.0057 247.578C20.0493 247.594 24.2922 249.167 26.616 246.424C26.7037 246.322 26.8557 246.31 26.9574 246.396C27.0596 246.482 27.0717 246.635 26.986 246.738C25.7919 248.147 24.1556 248.523 22.7494 248.523Z" fill="#FFB255"/>
-<path d="M22.8645 236.902C22.8543 236.902 22.8437 236.901 22.833 236.9C22.7008 236.883 22.6064 236.761 22.6233 236.629L22.7623 235.541L22.4495 235.048C22.3783 234.935 22.4117 234.785 22.5246 234.714C22.6364 234.642 22.7865 234.675 22.8587 234.789L23.2175 235.356C23.2475 235.404 23.2601 235.46 23.2528 235.516L23.1027 236.69C23.0887 236.813 22.9846 236.902 22.8645 236.902Z" fill="black"/>
-<path d="M19.3788 236.622C19.3671 236.622 19.355 236.621 19.3434 236.619C19.2112 236.6 19.1192 236.477 19.1386 236.345L19.296 235.274L18.911 234.769C18.8296 234.663 18.8505 234.511 18.957 234.43C19.0621 234.349 19.2141 234.369 19.296 234.475L19.7429 235.061C19.7826 235.113 19.7996 235.179 19.7904 235.243L19.6185 236.415C19.6005 236.535 19.4969 236.622 19.3788 236.622Z" fill="black"/>
-<path d="M18.5 232.649C18.5 232.649 20.5347 232.55 21.7439 232.649C22.953 232.748 22.953 232.748 22.953 232.748L23.5787 231.219V232.994C23.5787 232.994 24.8222 233.486 25.1858 233.882C25.5495 234.277 26.0618 235.263 26.0618 235.263L26.7213 234.408L26.3277 238.074C26.3277 238.074 26.947 236.386 28.1382 236.408C29.3294 236.431 29.7086 236.659 29.7086 236.659C29.7086 236.659 30.7216 233.142 30.5497 232.155C30.3778 231.169 29.4432 231.218 29.4432 231.218C29.4432 231.218 28.4612 227.865 23.63 227.865C18.7988 227.865 18.5 232.649 18.5 232.649Z" fill="black"/>
-<path d="M14.4165 251.066L14.1501 253.298L13.0025 262.954C12.7071 262.78 12.4214 262.601 12.1357 262.412C12.126 262.407 12.1212 262.402 12.1115 262.392C11.8597 262.228 11.6031 262.053 11.3561 261.874C11.2254 261.777 11.0946 261.685 10.9687 261.589C10.2133 261.017 9.49664 260.402 8.81871 259.739C8.71218 259.637 8.60564 259.531 8.50879 259.424C11.2738 254.417 14.4165 251.066 14.4165 251.066Z" fill="#2F2F2F"/>
-<path d="M21.5425 240.862C21.3173 240.862 21.1173 240.822 20.9778 240.772C20.8514 240.728 20.7861 240.589 20.8301 240.463C20.8752 240.338 21.0117 240.272 21.1386 240.316C21.173 240.328 21.9758 240.595 22.4591 239.931C22.5385 239.823 22.6901 239.8 22.7976 239.878C22.9056 239.957 22.9288 240.108 22.8504 240.216C22.4848 240.718 21.9686 240.862 21.5425 240.862Z" fill="black"/>
-<path d="M92.1045 266.277L92.1045 266.277L92.1124 266.275C97.4059 264.785 101.952 261.504 105.027 257.107C105.13 256.971 105.221 256.837 105.297 256.726L105.301 256.72C105.301 256.72 105.301 256.719 105.301 256.719C105.512 256.41 105.711 256.094 105.897 255.783C105.972 255.661 106.055 255.526 106.134 255.388C108.153 251.913 109.305 247.876 109.305 243.575C109.305 230.554 98.7517 220 85.7303 220C72.709 220 62.1553 230.554 62.1553 243.575C62.1553 247.69 63.212 251.559 65.0679 254.928C65.4058 255.543 65.7755 256.151 66.1778 256.737C70.4 263.019 77.5851 267.15 85.7303 267.15C87.8176 267.15 89.8455 266.877 91.7732 266.366L91.7801 266.364C91.8301 266.351 91.8819 266.337 91.9369 266.322C91.9908 266.308 92.0481 266.293 92.1045 266.277Z" fill="#85E0D3" stroke="white" stroke-width="2"/>
-<path d="M102.606 248.181C102.683 259.706 91.2119 257.769 91.2119 257.769L80.3505 253.474C80.3505 253.474 69.4698 247.721 76.6026 241.063C76.6026 241.063 74.1087 229.863 82.9412 227.616C82.9799 227.601 83.0187 227.591 83.0574 227.582C83.348 227.514 83.6288 227.456 83.9 227.417C92.0787 226.148 92.7809 237.155 92.7809 237.155C92.7809 237.155 102.543 238.961 102.606 248.181Z" fill="black"/>
-<path d="M104.218 256.521C101.274 260.738 96.9161 263.886 91.8413 265.314C91.7348 265.343 91.6234 265.373 91.5169 265.402C89.672 265.891 87.7302 266.152 85.7303 266.152C84.7957 266.152 83.8757 266.094 82.9701 265.987C82.2099 265.891 81.4593 265.76 80.7233 265.595C75.0287 264.307 70.1573 260.869 67.0049 256.177C70.0362 252.032 73.8374 248.894 76.4233 247.78C80.2487 246.134 86.3162 245.03 91.9188 246.633C91.9333 246.637 91.9527 246.637 91.9721 246.642C93.6572 247.126 95.2939 247.853 96.8193 248.884C100.117 251.112 102.519 253.698 104.218 256.521Z" fill="#FC7A82"/>
-<path d="M81.6133 242.694L81.6143 242.826L81.7247 245.297L81.8031 247.234C81.8031 247.234 82.1363 248.011 84.3918 247.865C87.4149 247.657 87.8904 246.032 87.8904 246.032L88.0192 241.49L87.1031 241.665L81.6133 242.694Z" fill="#FCC898"/>
-<path d="M81.6123 242.825L81.7227 245.297C83.1914 245.118 85.8649 244.482 86.9452 242.172C87.028 241.984 87.0832 241.815 87.1011 241.674L81.622 242.788L81.6123 242.825Z" fill="black"/>
-<path d="M80.6542 230.999C80.6542 230.999 77.8432 237.433 78.8557 241.047C79.8688 244.662 83.1621 244.093 85.1324 243.394C87.1028 242.696 88.0175 241.488 88.0175 241.488C88.0175 241.488 89.8581 242.112 91.089 240.4C92.3199 238.688 91.5621 236.723 90.5113 236.272C89.4605 235.822 88.2107 236.581 88.2107 236.581L89.1903 234.711C89.1903 234.711 85.069 225.208 80.6542 230.999Z" fill="#FCC898"/>
-<path d="M81.5267 230.046C81.5267 230.046 81.3233 233.47 84.7667 235.025C88.2106 236.58 90.5112 236.272 90.5112 236.272L90.2875 231.808L85.6117 227.797L82.1228 228.354L81.5267 230.046Z" fill="black"/>
-<path d="M87.9775 235.593L87.9906 237.542L89.7242 235.832L87.9775 235.593Z" fill="black"/>
-<path d="M105.041 255.265C104.862 255.565 104.673 255.865 104.475 256.156C104.392 256.277 104.31 256.398 104.218 256.519C101.274 260.736 96.9158 263.884 91.841 265.312C91.7345 265.341 91.6231 265.371 91.5166 265.4L91.8265 246.616C91.8265 246.616 91.8604 246.621 91.9185 246.631C91.933 246.635 91.9524 246.635 91.9718 246.64C93.0855 246.829 100.383 248.272 104.3 254.015C104.567 254.398 104.809 254.819 105.041 255.265Z" fill="#FC7A82"/>
-<path d="M79.8553 233.359C79.7521 233.359 79.6567 233.293 79.6243 233.189C79.5846 233.061 79.6558 232.926 79.7831 232.886L81.0654 232.485C81.1927 232.445 81.3288 232.516 81.369 232.644C81.4087 232.772 81.338 232.907 81.2102 232.947L79.9279 233.348C79.9037 233.356 79.8795 233.359 79.8553 233.359Z" fill="black"/>
-<path d="M88.3418 239.594C88.2764 239.594 88.2111 239.568 88.1636 239.516C88.0731 239.417 88.0794 239.264 88.1777 239.174L90.1611 237.348C90.2594 237.257 90.4129 237.264 90.5034 237.362C90.5935 237.461 90.5877 237.614 90.4889 237.704L88.5055 239.53C88.4595 239.573 88.4009 239.594 88.3418 239.594Z" fill="black"/>
-<path d="M90.351 239.354C90.3185 239.354 90.2856 239.347 90.2536 239.334L89.0697 238.814C88.9476 238.76 88.8915 238.618 88.9452 238.495C88.999 238.373 89.1413 238.317 89.2643 238.371L90.4483 238.89C90.5703 238.944 90.6265 239.087 90.5727 239.209C90.533 239.3 90.4439 239.354 90.351 239.354Z" fill="black"/>
-<path d="M80.2512 236.605C80.2348 236.605 80.2183 236.603 80.2019 236.6C80.0706 236.573 79.9868 236.444 80.014 236.313L80.2493 235.18L79.7878 234.648C79.7002 234.547 79.7108 234.394 79.812 234.306C79.9132 234.218 80.0663 234.229 80.1534 234.331L80.6938 234.953C80.7437 235.01 80.7636 235.087 80.7481 235.161L80.4885 236.412C80.4643 236.526 80.3636 236.605 80.2512 236.605Z" fill="black"/>
-<path d="M84.1305 236.802C84.114 236.802 84.0975 236.801 84.0811 236.797C83.9499 236.77 83.8661 236.642 83.8932 236.511L84.1281 235.377L83.6666 234.845C83.5789 234.745 83.5901 234.592 83.6908 234.503C83.792 234.416 83.9445 234.426 84.0327 234.528L84.5726 235.15C84.622 235.208 84.6418 235.284 84.6268 235.358L84.3673 236.609C84.3435 236.723 84.2428 236.802 84.1305 236.802Z" fill="black"/>
-<path d="M81.0176 239.43C80.9053 239.43 80.8051 239.351 80.7808 239.237C80.3489 237.174 81.3493 236.134 81.3919 236.091C81.4859 235.995 81.6389 235.994 81.7343 236.089C81.8292 236.183 81.8302 236.335 81.7367 236.431C81.696 236.473 80.8806 237.35 81.2544 239.138C81.282 239.269 81.1978 239.397 81.067 239.424C81.051 239.428 81.0341 239.43 81.0176 239.43Z" fill="black"/>
-<path d="M83.1093 241.409C82.9703 241.409 82.6381 241.345 81.9495 240.583C81.8672 240.492 81.8382 240.369 81.8701 240.254C81.904 240.134 81.9984 240.04 82.1234 240.003C83.779 239.507 83.9988 239.672 84.1039 239.752C84.1882 239.815 84.316 239.968 84.2225 240.283C84.1102 240.66 83.593 241.378 83.1514 241.408C83.1398 241.408 83.1257 241.409 83.1093 241.409ZM82.4517 240.411C82.8095 240.781 83.0473 240.923 83.1194 240.924C83.2899 240.901 83.7068 240.376 83.7645 240.135C83.6279 240.137 83.2294 240.188 82.4517 240.411Z" fill="black"/>
-<path d="M74.7949 255.158C74.7702 255.187 72.4575 257.666 72.8963 262.125C72.7573 262.035 72.6231 261.941 72.4895 261.842C72.4512 261.817 72.4178 261.787 72.3796 261.758C72.0953 257.331 74.4177 254.856 74.4473 254.823C74.5407 254.728 74.6908 254.726 74.7911 254.814C74.886 254.908 74.8884 255.063 74.7949 255.158Z" fill="white"/>
-<path d="M84.6905 248.274L82.9715 265.987C82.2112 265.891 81.4606 265.76 80.7246 265.595L83.078 248.681L84.6905 248.274Z" fill="white"/>
-<path d="M81.7608 246.23C81.7608 246.23 79.4651 251.31 80.0147 251.231C80.5638 251.151 83.9472 248.4 83.9472 248.4C83.9472 248.4 86.162 251.513 86.9344 251.571C87.7068 251.628 87.959 246 87.959 246C87.959 246 87.3746 247.592 84.491 247.862C81.6073 248.131 81.7608 246.23 81.7608 246.23Z" fill="white"/>
-<path d="M161.569 262.017L161.569 262.017L161.564 262.021C161.39 262.157 161.203 262.304 161.009 262.447C160.813 262.596 160.614 262.739 160.419 262.874C160.224 263.01 160.024 263.149 159.816 263.284L159.271 262.446L159.816 263.284C159.618 263.413 159.413 263.543 159.206 263.67C159.163 263.697 159.12 263.723 159.083 263.745L159.08 263.746C159.039 263.771 159.005 263.792 158.973 263.812L158.97 263.814C158.858 263.883 158.736 263.955 158.612 264.023C158.603 264.028 158.593 264.034 158.573 264.046L158.57 264.048C158.552 264.058 158.525 264.074 158.494 264.091C158.349 264.174 158.205 264.252 158.074 264.322C157.888 264.425 157.685 264.534 157.476 264.636C157.343 264.702 157.203 264.772 157.057 264.843L157.048 264.847L157.048 264.847C157.026 264.857 156.896 264.923 156.776 264.973C156.662 265.028 156.52 265.096 156.368 265.157C156.174 265.246 155.977 265.331 155.774 265.413L155.768 265.415C155.747 265.423 155.715 265.437 155.676 265.452C155.582 265.491 155.449 265.545 155.345 265.583C155.318 265.593 155.29 265.604 155.262 265.614C155.148 265.657 155.021 265.705 154.892 265.75C154.812 265.782 154.738 265.807 154.69 265.823L154.689 265.823C154.503 265.888 154.302 265.959 154.093 266.025C153.919 266.084 153.735 266.141 153.546 266.19C153.39 266.242 153.24 266.283 153.117 266.317L153.107 266.32L153.095 266.323L153.083 266.326C153.056 266.333 153.037 266.338 153.023 266.341C153.016 266.343 153.009 266.345 153.004 266.347L152.992 266.35L152.992 266.35C152.768 266.411 152.543 266.467 152.318 266.519C151.599 266.691 150.866 266.827 150.129 266.928L150.127 266.928C149.879 266.962 149.607 266.998 149.331 267.024C149.068 267.053 148.802 267.073 148.561 267.092L148.529 267.094L148.518 267.095L148.518 267.095C147.977 267.13 147.432 267.15 146.88 267.15C145.985 267.15 145.095 267.1 144.226 267.005C144.092 266.993 143.966 266.977 143.862 266.964L143.855 266.963L143.687 266.941L143.681 266.938C142.617 266.791 141.575 266.576 140.559 266.29C139.309 265.946 138.101 265.496 136.951 264.962C136.772 264.88 136.594 264.793 136.429 264.713L136.424 264.711L136.418 264.708C135.28 264.146 134.197 263.494 133.179 262.76C132.159 262.032 131.198 261.223 130.308 260.338M161.569 262.017L130.306 260.337C130.307 260.337 130.307 260.338 130.308 260.338M161.569 262.017L161.576 262.011C162.334 261.405 163.05 260.755 163.724 260.071C163.725 260.07 163.726 260.069 163.727 260.068L163.727 260.068C163.768 260.026 163.828 259.967 163.887 259.903L163.93 259.86C163.93 259.86 163.93 259.86 163.93 259.86C163.999 259.791 164.062 259.722 164.123 259.65C164.435 259.313 164.744 258.964 165.042 258.596C166.967 256.281 168.452 253.587 169.377 250.645C170.081 248.412 170.455 246.035 170.455 243.575C170.455 230.554 159.901 220 146.88 220C133.858 220 123.305 230.554 123.305 243.575C123.305 247.441 124.235 251.088 125.889 254.311C126.263 255.043 126.677 255.756 127.133 256.445C127.317 256.732 127.51 257.006 127.69 257.261L127.695 257.269L127.699 257.274C128.176 257.942 128.686 258.59 129.246 259.211M161.569 262.017L129.246 259.211M130.308 260.338C129.94 259.976 129.586 259.602 129.246 259.211M130.308 260.338C130.308 260.338 130.309 260.339 130.309 260.339L129.246 259.211M143.96 266.966C143.969 266.966 143.963 266.966 143.947 266.966H143.96Z" fill="#C0A2FC" stroke="white" stroke-width="2"/>
-<path d="M142.009 234.223C142.009 234.223 134.3 237.171 133.907 245.895C133.339 258.506 148.452 255.583 148.452 255.583C148.452 255.583 157.012 256.471 159.379 250.161C161.747 243.851 154.136 240.042 154.136 240.042C154.136 240.042 144.909 231.915 142.009 234.223Z" fill="black"/>
-<path d="M164.269 257.962C162.318 252.727 156.764 244.766 147.505 245.105C146.876 245.129 146.265 245.188 145.68 245.275C145.66 245.28 145.641 245.284 145.621 245.284C137.433 246.568 132.014 253.894 129.995 258.548C130.32 258.92 130.659 259.279 131.012 259.627C131.864 260.475 132.784 261.25 133.762 261.947C134.736 262.649 135.772 263.274 136.862 263.811C137.031 263.894 137.201 263.976 137.37 264.053C138.474 264.567 139.631 264.998 140.827 265.327C141.844 265.613 142.89 265.826 143.96 265.966C143.97 265.966 143.975 265.966 143.985 265.971C144.096 265.985 144.207 266 144.319 266.01C145.161 266.102 146.018 266.15 146.88 266.15C147.408 266.15 147.931 266.131 148.454 266.097C148.711 266.077 148.972 266.058 149.229 266.029C149.486 266.005 149.742 265.971 149.994 265.937C150.701 265.84 151.403 265.709 152.091 265.545C152.304 265.496 152.517 265.443 152.73 265.385C152.764 265.375 152.803 265.366 152.841 265.356C152.982 265.317 153.122 265.278 153.263 265.23C153.437 265.186 153.611 265.133 153.781 265.075C153.979 265.012 154.173 264.944 154.367 264.877C154.425 264.857 154.483 264.838 154.541 264.814C154.696 264.76 154.846 264.702 155.001 264.644C155.137 264.596 155.267 264.538 155.403 264.484C155.597 264.407 155.786 264.325 155.974 264.237C156.11 264.184 156.241 264.121 156.371 264.058C156.459 264.024 156.541 263.981 156.623 263.942C156.764 263.874 156.899 263.806 157.035 263.739C157.224 263.647 157.413 263.545 157.597 263.443C157.732 263.371 157.868 263.298 158.003 263.22C158.047 263.196 158.086 263.172 158.129 263.148C158.236 263.09 158.342 263.027 158.444 262.964C158.521 262.915 158.604 262.867 158.681 262.819C158.88 262.697 159.078 262.572 159.272 262.446C159.466 262.32 159.655 262.189 159.848 262.053C160.037 261.923 160.226 261.787 160.41 261.647C160.594 261.511 160.773 261.371 160.952 261.23C161.679 260.649 162.366 260.025 163.015 259.366L164.269 257.962Z" fill="white"/>
-<path d="M158.444 262.963C155.064 264.987 151.107 266.149 146.88 266.149C146.018 266.149 145.161 266.101 144.323 266.009C144.207 265.999 144.096 265.985 143.984 265.97C143.975 265.965 143.97 265.965 143.96 265.965C142.89 265.825 141.844 265.612 140.827 265.326C139.631 264.997 138.474 264.566 137.37 264.052C137.2 263.975 137.031 263.893 136.861 263.81C135.772 263.273 134.735 262.648 133.762 261.946C133.719 258.619 133.694 255.995 133.694 255.995L142.294 249.612L146.517 246.479C146.57 246.479 146.623 246.479 146.677 246.484C149.001 246.533 150.962 247.007 152.507 247.579C152.754 247.676 152.991 247.768 153.219 247.864C155.364 248.775 156.526 249.797 156.526 249.797C156.526 249.797 157.533 256.469 158.444 262.963Z" fill="white"/>
-<path d="M145.483 241.98L145.662 246.749C145.662 246.749 146.555 248.214 149.537 248.344C151.758 248.442 152.547 247.81 152.547 247.81L152.202 245.625L151.804 243.126L146.392 242.146L145.483 241.98Z" fill="#F2AC82"/>
-<path d="M145.483 241.738C145.501 241.877 145.56 242.039 145.649 242.223C146.881 244.719 150.919 245.559 152.202 245.625L151.804 243.127L145.483 241.738Z" fill="black"/>
-<path d="M152.505 231.501C152.505 231.501 155.41 237.773 154.491 241.352C153.572 244.931 150.318 244.442 148.364 243.797C146.41 243.151 145.483 241.982 145.483 241.982C145.483 241.982 143.685 242.636 142.437 240.978C141.188 239.32 141.892 237.369 142.916 236.903C143.941 236.437 145.188 237.157 145.188 237.157L144.183 235.338C144.183 235.338 148.036 225.896 152.505 231.501Z" fill="#F2AC82"/>
-<path d="M153.442 234.107C153.431 234.107 153.419 234.106 153.407 234.104L152.167 233.92C152.034 233.901 151.943 233.778 151.963 233.645C151.983 233.513 152.106 233.425 152.238 233.442L153.478 233.626C153.61 233.645 153.702 233.768 153.682 233.901C153.664 234.021 153.56 234.107 153.442 234.107Z" fill="black"/>
-<path d="M148.817 234.62C148.723 234.62 148.633 234.564 148.594 234.471C148.542 234.348 148.6 234.206 148.723 234.154L150.163 233.55C150.286 233.499 150.428 233.556 150.48 233.679C150.532 233.803 150.473 233.945 150.35 233.996L148.911 234.601C148.881 234.613 148.849 234.62 148.817 234.62Z" fill="black"/>
-<path d="M152.467 239.25C152.459 239.25 152.451 239.25 152.443 239.249C152.309 239.236 152.213 239.117 152.226 238.984C152.425 237.026 151.614 236.041 151.605 236.031C151.519 235.929 151.531 235.777 151.633 235.69C151.734 235.604 151.886 235.615 151.973 235.716C152.013 235.761 152.929 236.853 152.708 239.033C152.695 239.157 152.59 239.25 152.467 239.25Z" fill="black"/>
-<path d="M150.489 241.37C150.394 241.37 150.294 241.357 150.191 241.328C149.39 241.099 149.3 240.181 149.297 240.142C149.29 240.064 149.322 239.987 149.382 239.937C149.442 239.886 149.523 239.867 149.599 239.887L151.479 240.374C151.558 240.395 151.621 240.453 151.647 240.529C151.674 240.606 151.66 240.691 151.611 240.755C151.589 240.783 151.134 241.37 150.489 241.37ZM149.883 240.461C149.961 240.625 150.095 240.797 150.324 240.862C150.553 240.929 150.768 240.841 150.93 240.732L149.883 240.461Z" fill="black"/>
-<path d="M145.196 239.773C145.163 239.773 145.129 239.766 145.097 239.751L142.543 238.59C142.422 238.535 142.368 238.391 142.423 238.27C142.478 238.149 142.621 238.094 142.744 238.15L145.297 239.31C145.419 239.366 145.472 239.509 145.417 239.631C145.376 239.72 145.289 239.773 145.196 239.773Z" fill="black"/>
-<path d="M142.921 240.363C142.858 240.363 142.796 240.339 142.749 240.291C142.655 240.197 142.656 240.043 142.751 239.949L143.876 238.836C143.972 238.742 144.125 238.743 144.219 238.838C144.313 238.933 144.312 239.087 144.217 239.181L143.091 240.293C143.043 240.34 142.982 240.363 142.921 240.363Z" fill="black"/>
-<path d="M148.817 232.18C148.817 232.18 150.272 231.972 151.789 232.835C153.306 233.699 153.618 234.54 153.618 234.54C153.618 234.54 154.556 232.311 152.499 230.344C150.442 228.377 146.496 227.832 144.221 230.159C141.946 232.486 142.086 234.54 142.086 234.54L145.188 237.157C145.188 237.157 149.978 236.288 148.817 232.18Z" fill="black"/>
-<path d="M143.667 230.784C143.667 230.784 140.74 225.225 137.773 227.575C134.807 229.924 138.11 234.664 141.057 235.01C144.004 235.356 143.667 230.784 143.667 230.784Z" fill="black"/>
-<path d="M144.323 266.009C144.207 265.999 144.096 265.985 143.984 265.97C143.975 265.965 143.97 265.965 143.96 265.965C142.89 265.825 141.844 265.612 140.827 265.326L142.294 249.613L142.343 249.114L145.484 247.534L144.918 246.162C144.918 246.162 148.25 259.259 144.323 266.009Z" fill="white"/>
-<path d="M158.681 262.818C158.603 262.866 158.521 262.914 158.443 262.963C158.342 263.026 158.235 263.089 158.129 263.147C159.179 257.661 157.029 254.518 157.005 254.484C156.928 254.373 156.947 254.213 157.054 254.13C157.165 254.048 157.315 254.072 157.393 254.184C157.489 254.319 159.591 257.37 158.681 262.818Z" fill="#92B4EA"/>
-<path d="M149.809 248.587C147.059 248.587 145.503 246.935 145.482 246.913C145.392 246.815 145.399 246.662 145.497 246.571C145.596 246.481 145.749 246.487 145.839 246.585C145.941 246.696 148.394 249.275 152.729 247.479C152.852 247.429 152.994 247.486 153.045 247.61C153.096 247.734 153.038 247.875 152.914 247.926C151.757 248.406 150.717 248.587 149.809 248.587Z" fill="#F7B369"/>
-<path d="M145.828 241.738C145.828 242.241 145.42 242.648 144.918 242.648C144.415 242.648 144.008 242.241 144.008 241.738C144.008 241.236 144.415 240.828 144.918 240.828C145.42 240.828 145.828 241.236 145.828 241.738Z" fill="#E6F4FF"/>
-<path d="M149.538 237.229C149.404 237.229 149.296 237.121 149.296 236.987V235.874C149.296 235.81 149.321 235.749 149.366 235.704L149.859 235.204C149.953 235.108 150.106 235.107 150.201 235.201C150.296 235.295 150.298 235.449 150.204 235.544L149.781 235.973V236.987C149.78 237.121 149.672 237.229 149.538 237.229Z" fill="black"/>
-<path d="M153.196 237.229C153.063 237.229 152.954 237.121 152.954 236.987V235.874C152.954 235.81 152.979 235.749 153.024 235.704L153.517 235.204C153.61 235.108 153.763 235.107 153.859 235.201C153.955 235.295 153.955 235.448 153.862 235.543L153.438 235.973V236.987C153.438 237.121 153.33 237.229 153.196 237.229Z" fill="black"/>
-<path d="M195.039 263.246C196.373 264.131 197.809 264.884 199.316 265.481C200.17 265.824 201.052 266.117 201.95 266.354C202.057 266.385 202.168 266.411 202.251 266.43C202.263 266.433 202.275 266.436 202.286 266.439C202.845 266.579 203.409 266.7 203.983 266.801L203.984 266.801C204.502 266.892 205.028 266.968 205.562 267.024M195.039 263.246C194.625 262.974 194.216 262.685 193.82 262.386C193.032 261.79 192.283 261.148 191.575 260.455M195.039 263.246C195.039 263.246 195.04 263.247 195.04 263.247L195.59 262.412L195.038 263.245C195.038 263.246 195.039 263.246 195.039 263.246ZM191.575 260.455C189.891 258.816 188.46 256.93 187.328 254.858L187.328 254.858C185.497 251.504 184.455 247.66 184.455 243.575C184.455 230.553 195.014 220 208.035 220C221.056 220 231.61 230.554 231.61 243.575C231.61 246.738 230.988 249.758 229.855 252.517C229.025 254.535 227.923 256.416 226.593 258.11C223.54 262.006 219.283 264.919 214.377 266.285C212.356 266.852 210.227 267.15 208.035 267.15C207.204 267.15 206.379 267.11 205.562 267.024M191.575 260.455C191.575 260.455 191.575 260.455 191.576 260.456L192.273 259.739L191.574 260.454C191.574 260.454 191.574 260.454 191.575 260.455ZM205.562 267.024L205.667 266.029M205.562 267.024C205.562 267.024 205.563 267.024 205.563 267.024L205.667 266.029M205.667 266.029C205.159 265.976 204.655 265.903 204.156 265.816L225.806 257.492C222.882 261.225 218.804 264.015 214.107 265.322C212.175 265.864 210.137 266.15 208.035 266.15C207.236 266.15 206.447 266.111 205.667 266.029Z" fill="#B0E6FD" stroke="white" stroke-width="2"/>
-<path d="M202.474 233.306C202.474 233.306 200.314 235.118 200.11 236.936C199.906 238.754 199.906 238.754 199.906 238.754C199.906 238.754 197.317 239.582 196.535 242.71C195.474 246.957 198.368 248.434 198.368 248.434L220.854 250.862C220.854 250.862 223.497 248.546 222.968 244.824C222.439 241.103 220.261 239.919 220.261 239.919C220.261 239.919 220.619 237.526 219.415 235.122C218.026 232.349 215.173 232.321 215.173 232.321L202.474 233.306Z" fill="black"/>
-<path d="M225.806 257.492C222.881 261.226 218.804 264.015 214.107 265.322C212.175 265.864 210.136 266.15 208.035 266.15C207.236 266.15 206.446 266.111 205.667 266.029C205.158 265.976 204.655 265.903 204.156 265.816C203.604 265.719 203.062 265.603 202.524 265.467C202.422 265.443 202.316 265.419 202.214 265.39C201.352 265.162 200.505 264.881 199.686 264.552C198.243 263.981 196.868 263.259 195.59 262.412C195.638 261.206 195.721 260.523 195.721 260.523L199.367 249.023L199.725 247.89C199.938 247.212 200.374 246.645 200.95 246.267C201.246 246.069 201.58 245.924 201.938 245.846C202.606 245.691 203.439 245.522 204.384 245.362C204.679 245.318 204.974 245.27 205.289 245.226C206.267 245.091 207.342 244.984 208.451 244.936C209.119 244.907 209.788 244.902 210.461 244.926C210.761 244.936 211.061 244.95 211.361 244.97C211.71 244.999 212.054 245.033 212.403 245.071C212.79 245.12 213.172 245.178 213.555 245.255C214.427 245.42 215.279 245.652 216.088 245.967C218.233 246.805 220.097 248.529 221.632 250.461C223.588 252.926 225.012 255.734 225.806 257.492Z" fill="#F6FF82"/>
-<path d="M211.431 243.892C211.431 243.892 213.012 244.41 213.491 245.146C214.029 245.971 213.024 248.143 210.164 249.832C208.074 251.066 206.311 251.502 206.311 251.502C206.311 251.502 204.491 248.986 204.142 247.456C203.794 245.926 205.074 243.92 205.074 243.92L211.431 243.892Z" fill="#FC7A82"/>
-<path d="M216.548 236.85C214.926 243.75 208.868 241.688 208.868 241.688C208.868 241.688 199.552 240.438 203.058 236.133C203.058 236.133 201.159 233.228 201.174 231.804C201.184 230.821 203.251 228.531 206.162 227.664C206.452 227.577 206.747 227.504 207.053 227.451C208.128 227.253 209.295 227.262 210.495 227.601C215.463 228.996 217.177 234.167 216.548 236.85Z" fill="black"/>
-<path d="M205.057 243.31L205.073 243.436L205.467 245.8L205.768 247.654C205.768 247.654 205.662 248.107 207.814 247.703C210.697 247.15 211.482 245.788 211.482 245.788L211.076 241.404L210.215 241.679L205.057 243.31Z" fill="#A95832"/>
-<path d="M205.072 243.436L205.466 245.8C206.858 245.457 209.355 244.533 210.124 242.185C210.182 241.995 210.215 241.826 210.216 241.688L205.077 243.399L205.072 243.436Z" fill="black"/>
-<path d="M202.771 232.174C202.771 232.174 200.818 238.69 202.214 242.048C203.61 245.406 206.71 244.474 208.524 243.572C210.337 242.67 211.076 241.401 211.076 241.401C211.076 241.401 212.919 241.787 213.903 239.997C214.888 238.207 213.929 236.405 212.866 236.095C211.803 235.785 210.689 236.66 210.689 236.66L211.413 234.747C211.413 234.747 206.34 226.09 202.771 232.174Z" fill="#A95832"/>
-<path d="M203.783 239.884C203.691 239.884 203.604 239.831 203.563 239.742C202.976 238.453 203.745 236.971 203.777 236.909C203.84 236.79 203.987 236.745 204.104 236.807C204.222 236.87 204.268 237.016 204.206 237.134C204.199 237.147 203.516 238.471 204.004 239.541C204.059 239.663 204.005 239.807 203.883 239.862C203.851 239.877 203.817 239.884 203.783 239.884Z" fill="black"/>
-<path d="M211.3 239.571C211.25 239.571 211.2 239.556 211.157 239.525C211.049 239.446 211.026 239.295 211.104 239.187L212.555 237.192C212.634 237.083 212.786 237.06 212.893 237.139C213.002 237.217 213.025 237.369 212.947 237.477L211.495 239.472C211.448 239.537 211.375 239.571 211.3 239.571Z" fill="black"/>
-<path d="M213.017 238.867C213.011 238.867 211.827 238.84 211.827 238.84C211.694 238.837 211.588 238.726 211.591 238.592C211.595 238.458 211.709 238.363 211.839 238.355L213.023 238.383C213.156 238.385 213.262 238.496 213.259 238.631C213.256 238.762 213.149 238.867 213.017 238.867Z" fill="black"/>
-<path d="M210.691 236.66C210.691 236.66 207.815 235.623 206.59 233.805C205.365 231.986 205.204 231.016 205.204 231.016C205.204 231.016 204.972 232.342 204.026 232.767C203.081 233.194 202.474 233.306 202.474 233.306L202.078 235.198L201.934 230.826C201.934 230.826 204.417 229.06 204.774 229.014C205.131 228.968 209.737 229.198 210.032 229.436C210.327 229.674 212.194 232.989 212.194 232.989L212.495 234.556L210.691 236.66Z" fill="black"/>
-<path d="M211.291 244.202C211.291 244.202 208.213 247.104 206.411 247.12C204.608 247.137 204.861 245.638 204.861 245.638L205.158 249.097L206.75 249.619L210.731 248.015L212.965 245.514L211.291 244.202Z" fill="#FC7A82"/>
-<path d="M208.037 249.794C208.037 249.794 209.387 251.64 209.661 254.537C209.936 257.434 208.253 260.69 208.253 260.69C208.253 260.69 205.557 257.632 205.2 255.324C204.843 253.015 206.078 249.738 206.078 249.738C206.078 249.738 204.856 252.518 204.554 254.22C204.252 255.922 203.835 258.108 203.835 258.108C203.835 258.108 202.198 255.939 202.296 253.364C202.394 250.789 205.16 249.097 205.16 249.097L208.037 249.794Z" fill="#FC7A82"/>
-<path d="M207.691 250.644C206.946 250.644 205.987 250.384 205.852 250.347C205.838 250.343 205.826 250.338 205.815 250.334C205.763 250.309 205.718 250.266 205.694 250.208C205.531 249.824 204.743 247.868 205.251 247.165C205.35 247.029 205.542 246.875 205.884 246.9C206.159 246.923 206.376 247.052 206.511 247.275C206.915 247.939 206.465 249.348 206.242 249.947C206.765 250.073 207.656 250.24 208.068 250.12C208.503 249.995 208.75 248.495 208.576 247.648C208.549 247.517 208.633 247.389 208.765 247.363C208.893 247.337 209.024 247.42 209.051 247.551C209.207 248.315 209.104 250.325 208.202 250.585C208.057 250.627 207.881 250.644 207.691 250.644ZM205.804 247.381C205.693 247.381 205.661 247.426 205.644 247.448C205.441 247.729 205.642 248.633 205.912 249.421C206.138 248.717 206.307 247.87 206.097 247.526C206.044 247.437 205.966 247.393 205.845 247.383C205.83 247.382 205.817 247.381 205.804 247.381Z" fill="#F6FF82"/>
-<path d="M207.605 254.569C207.397 254.569 207.216 254.481 207.082 254.314C206.439 253.514 207.037 250.755 207.107 250.444C207.131 250.338 207.221 250.262 207.328 250.255C207.433 250.25 207.534 250.313 207.571 250.414C207.757 250.927 208.66 253.523 208.113 254.308C207.994 254.478 207.82 254.569 207.608 254.569C207.607 254.569 207.605 254.569 207.605 254.569ZM207.406 251.47C207.249 252.49 207.176 253.659 207.459 254.011C207.502 254.064 207.539 254.071 207.606 254.085C207.666 254.085 207.691 254.065 207.715 254.031C207.944 253.702 207.718 252.529 207.406 251.47Z" fill="#F6FF82"/>
-<path d="M202.525 265.467C202.423 265.443 202.316 265.419 202.215 265.39C201.353 265.162 200.505 264.881 199.687 264.552C198.244 263.981 196.869 263.259 195.59 262.412C195.193 262.15 194.801 261.874 194.423 261.589C193.668 261.017 192.951 260.402 192.273 259.739C192.283 259.676 192.298 259.613 192.312 259.55C193.547 253.962 195.891 250.524 197.828 248.529C199.581 246.727 201.259 246.088 201.259 246.088C201.24 246.146 200.97 246.204 200.951 246.267C200.762 246.965 200.675 247.957 200.67 249.163C200.656 253.323 201.605 259.986 202.525 265.467Z" fill="#F6FF82"/>
-<path d="M205.971 237.519C205.822 237.506 205.729 237.4 205.729 237.277C205.729 237.254 205.843 236.065 205.843 236.065C205.856 235.932 205.976 235.833 206.107 235.847C206.233 235.859 206.326 235.964 206.326 236.088C206.325 236.111 206.212 237.3 206.212 237.3C206.2 237.425 206.094 237.519 205.971 237.519Z" fill="black"/>
-<path d="M202.474 237.426C202.464 237.426 202.453 237.425 202.443 237.424C202.321 237.408 202.231 237.304 202.231 237.185C202.231 237.174 202.232 237.164 202.233 237.153L202.387 235.969C202.405 235.837 202.527 235.743 202.659 235.761C202.781 235.776 202.87 235.88 202.87 236C202.87 236.01 202.869 236.021 202.868 236.032L202.714 237.216C202.698 237.338 202.594 237.426 202.474 237.426Z" fill="black"/>
-<path d="M205.715 241.629C205.16 241.629 204.641 241.331 204.616 241.316C204.538 241.271 204.492 241.185 204.496 241.094C204.501 241.004 204.556 240.923 204.638 240.886L206.36 240.105C206.436 240.07 206.524 240.077 206.594 240.124C206.663 240.17 206.704 240.248 206.702 240.332C206.699 240.452 206.648 241.506 205.921 241.613C205.852 241.625 205.784 241.629 205.715 241.629ZM205.353 241.094C205.508 241.135 205.684 241.159 205.849 241.136C205.999 241.113 206.095 240.931 206.152 240.732L205.353 241.094Z" fill="black"/>
-<path d="M202.126 235.214C202.015 235.214 201.915 235.137 201.89 235.024C201.861 234.894 201.942 234.764 202.073 234.735L203.152 234.494C203.285 234.465 203.412 234.547 203.442 234.677C203.471 234.808 203.389 234.937 203.258 234.966L202.179 235.208C202.161 235.212 202.143 235.214 202.126 235.214Z" fill="black"/>
-<path d="M207.026 235.456C207.005 235.456 206.982 235.453 206.96 235.447L205.257 234.963C205.128 234.926 205.054 234.793 205.09 234.664C205.127 234.535 205.258 234.459 205.389 234.497L207.093 234.981C207.222 235.018 207.296 235.152 207.259 235.281C207.229 235.387 207.132 235.456 207.026 235.456Z" fill="black"/>
-<path d="M206.084 236.33C206.032 236.33 205.978 236.312 205.934 236.278L205.535 235.96C205.43 235.877 205.413 235.725 205.496 235.62C205.58 235.515 205.731 235.499 205.836 235.581L206.235 235.898C206.34 235.981 206.357 236.134 206.274 236.238C206.226 236.298 206.155 236.33 206.084 236.33Z" fill="black"/>
-<path d="M202.628 236.242C202.568 236.242 202.509 236.221 202.462 236.177L202.13 235.866C202.033 235.775 202.028 235.621 202.119 235.523C202.211 235.425 202.364 235.421 202.461 235.512L202.794 235.823C202.891 235.914 202.897 236.068 202.805 236.166C202.757 236.217 202.693 236.242 202.628 236.242Z" fill="black"/>
-<path d="M30.9492 319.429L30.9492 319.429L30.9571 319.427C36.2507 317.937 40.7972 314.656 43.872 310.259C43.9748 310.123 44.0654 309.99 44.1415 309.878L44.1455 309.872C44.1456 309.872 44.1458 309.871 44.146 309.871C44.3571 309.562 44.5556 309.246 44.7415 308.935C44.8164 308.813 44.8997 308.678 44.9791 308.54C46.9973 305.065 48.1501 301.029 48.1501 296.727C48.1501 283.706 37.5964 273.152 24.5751 273.152C11.5537 273.152 1 283.706 1 296.727C1 300.842 2.05671 304.711 3.91268 308.08C4.25048 308.695 4.62024 309.303 5.02251 309.889C9.24469 316.171 16.4299 320.302 24.5751 320.302C26.6623 320.302 28.6902 320.029 30.6179 319.518L30.6248 319.516C30.6749 319.503 30.7266 319.489 30.7816 319.474L30.7828 319.474C30.8364 319.46 30.8932 319.445 30.9492 319.429Z" fill="#F6FF82" stroke="white" stroke-width="2"/>
-<path d="M20.8899 289.916C20.8899 289.916 17.0649 288.689 17.0441 291.564C17.0198 294.895 20.9673 294.558 20.9673 294.558L20.8899 289.916Z" fill="#FBD7C6"/>
-<path d="M19.314 293.083C19.2384 293.083 19.1644 293.048 19.1169 292.982L17.6928 290.992C17.6148 290.883 17.64 290.732 17.7489 290.654C17.8584 290.576 18.009 290.602 18.0869 290.711L19.5111 292.7C19.589 292.809 19.5638 292.96 19.4549 293.038C19.4123 293.069 19.3629 293.083 19.314 293.083Z" fill="black"/>
-<path d="M17.8261 292.586C17.7041 292.586 17.5995 292.494 17.5854 292.37C17.5709 292.238 17.6668 292.118 17.7995 292.103L18.8474 291.987C18.9839 291.976 19.0996 292.069 19.1147 292.201C19.1292 292.334 19.0333 292.454 18.9006 292.469L17.8527 292.584C17.8435 292.585 17.8348 292.586 17.8261 292.586Z" fill="black"/>
-<path d="M46.3944 302.541C43.8377 312.202 35.0391 319.315 24.5748 319.315C17.534 319.315 11.2486 316.095 7.1084 311.049C13.7134 302.968 20.8365 302.106 20.8365 302.106C20.8365 302.106 20.667 300.885 20.8365 300.556C20.9575 300.324 21.403 300.111 21.65 300.009C21.7468 299.97 21.8098 299.946 21.8098 299.946C21.8098 299.946 22.1681 299.883 22.7298 299.796C23.8581 299.617 25.8047 299.331 27.3156 299.22C27.7756 299.181 28.1969 299.161 28.5407 299.166C28.9135 299.171 29.1992 299.205 29.3542 299.278C29.543 299.365 29.7609 299.946 29.7609 299.946C29.7609 299.946 32.555 299.515 36.4143 299.946C38.037 300.13 42.0702 300.876 46.3944 302.541Z" fill="#2F2F2F"/>
-<path d="M28.2528 297.145L28.6441 301.004C28.6441 301.004 27.2277 301.947 25.3663 302.152C22.8536 302.429 21.8653 301.016 21.8653 301.016L21.4619 296.485L28.2528 297.145Z" fill="#FBD7C6"/>
-<path d="M28.2526 297.139C28.2526 297.139 27.3384 300.378 24.4417 300.218C22.1067 300.088 21.1436 296.884 21.1436 296.884L28.2526 297.139Z" fill="black"/>
-<path d="M19.425 286.866C19.425 286.866 18.4711 294.086 20.7881 297.028C23.2732 300.184 26.6769 298.855 27.8919 297.726C29.531 296.204 29.7169 294.287 29.7169 294.287C29.7169 294.287 32.189 294.544 32.8136 292.489C33.4698 290.33 31.8021 288.794 30.5784 288.808C29.3552 288.822 29.2342 289.056 29.2342 289.056L29.3833 286.803C29.3833 286.803 21.2385 279.289 19.425 286.866Z" fill="#FBD7C6"/>
-<path d="M29.9102 292.248C29.8497 292.248 29.7892 292.225 29.7417 292.179C29.6459 292.086 29.6429 291.933 29.7359 291.837L31.4593 290.056C31.5523 289.959 31.7053 289.958 31.8017 290.05C31.8975 290.143 31.9005 290.297 31.8075 290.392L30.0841 292.173C30.0371 292.223 29.9737 292.248 29.9102 292.248Z" fill="black"/>
-<path d="M31.6682 292.382C31.6251 292.382 31.581 292.37 31.5413 292.346L30.3583 291.618C30.2445 291.548 30.2092 291.399 30.2789 291.285C30.3482 291.171 30.4973 291.135 30.6121 291.205L31.795 291.934C31.9088 292.004 31.9442 292.153 31.8745 292.267C31.8289 292.341 31.7495 292.382 31.6682 292.382Z" fill="black"/>
-<path d="M26.0779 290.636C25.9481 290.636 25.8406 290.533 25.8362 290.402L25.7951 289.206C25.7902 289.072 25.8948 288.96 26.0285 288.955C26.1592 288.949 26.274 289.055 26.2788 289.189L26.3205 290.385C26.3253 290.518 26.2207 290.63 26.0866 290.635C26.0837 290.636 26.0808 290.636 26.0779 290.636Z" fill="black"/>
-<path d="M21.5007 290.81C21.371 290.81 21.2635 290.708 21.2591 290.576L21.2179 289.381C21.2131 289.247 21.3177 289.135 21.4513 289.131C21.5821 289.115 21.6968 289.23 21.7017 289.364L21.7428 290.559C21.7477 290.693 21.6431 290.805 21.5094 290.81C21.5061 290.81 21.5031 290.81 21.5007 290.81Z" fill="black"/>
-<path d="M23.8961 293.77C23.7915 293.77 23.6951 293.702 23.6641 293.597C23.1169 291.748 23.575 290.163 23.5949 290.096C23.6331 289.969 23.7673 289.896 23.8956 289.933C24.0239 289.971 24.097 290.106 24.0593 290.234C24.0549 290.249 23.6234 291.754 24.1285 293.459C24.1663 293.587 24.0931 293.722 23.9653 293.76C23.9421 293.766 23.9188 293.77 23.8961 293.77Z" fill="black"/>
-<path d="M20.6874 288.37C20.6273 288.37 20.5668 288.347 20.5203 288.303C20.4235 288.21 20.4201 288.057 20.5121 287.961C20.9905 287.46 21.8539 287.387 21.8907 287.384C22.0229 287.373 22.1401 287.473 22.1507 287.607C22.1614 287.74 22.0616 287.856 21.929 287.867C21.9217 287.868 21.2123 287.929 20.8627 288.295C20.8147 288.345 20.7508 288.37 20.6874 288.37Z" fill="black"/>
-<path d="M26.5677 288.272C26.5265 288.272 26.4844 288.261 26.4461 288.239C26.0098 287.984 25.3101 288.121 25.3028 288.123C25.174 288.149 25.0442 288.065 25.0176 287.934C24.991 287.803 25.0752 287.675 25.206 287.649C25.2418 287.641 26.0916 287.472 26.6902 287.821C26.8054 287.888 26.8446 288.036 26.7773 288.152C26.7318 288.229 26.6509 288.272 26.5677 288.272Z" fill="black"/>
-<path d="M29.7788 289.826C29.7788 289.826 29.4456 292.122 29.1502 292.089C28.8548 292.055 28.6626 288.575 29.4456 288.28C30.2281 287.985 29.7788 289.826 29.7788 289.826Z" fill="black"/>
-<path d="M24.2487 295.81C24.1194 295.81 24.0128 295.708 24.007 295.578C24.0017 295.445 24.1058 295.332 24.239 295.326C24.2874 295.324 25.4297 295.264 25.8752 294.286C25.9309 294.164 26.0747 294.11 26.1958 294.165C26.3173 294.221 26.3711 294.365 26.3159 294.486C25.745 295.74 24.3179 295.808 24.2579 295.81C24.255 295.81 24.2516 295.81 24.2487 295.81Z" fill="black"/>
-<path d="M29.419 288.726C29.419 288.726 26.9553 288.628 25.2541 287.426C23.8174 286.413 23.2126 284.901 23.2126 284.901C23.2126 284.901 22.1512 284.812 20.6873 285.803C19.223 286.795 19.2302 289.658 19.2302 289.658C19.2302 289.658 16.2386 284.936 17.4875 283.164C18.7363 281.392 20.9662 282.721 20.9662 282.721C20.9662 282.721 20.4442 279.571 24.8813 280.605C29.3183 281.638 30.4858 285.379 30.4858 285.379C30.4858 285.379 31.9763 285.182 32.2223 286.56C32.4683 287.938 31.9652 289.372 31.9652 289.372C31.9652 289.372 31.2872 288.783 30.3028 289.197C29.3183 289.612 29.3183 289.612 29.3183 289.612L29.419 288.726Z" fill="black"/>
-<path d="M21.8102 299.948C21.8102 299.948 22.6905 300.783 24.7272 300.834C26.7639 300.885 28.5502 299.239 28.5502 299.239L29.8625 301.647L26.6177 303.628L21.1797 303.17V300.732L21.8102 299.948Z" fill="#2F2F2F"/>
-<path d="M24.0568 303.875C21.8559 303.875 20.655 302.266 20.641 302.246C20.5625 302.138 20.5862 301.987 20.6942 301.909C20.8022 301.831 20.9538 301.853 21.0322 301.961C21.0942 302.046 22.5924 304.034 25.3913 303.182C25.5201 303.143 25.6543 303.215 25.6935 303.343C25.7322 303.471 25.6601 303.606 25.5322 303.645C24.9996 303.807 24.5071 303.875 24.0568 303.875Z" fill="#2F2F2F"/>
-<path d="M107.448 306.254L107.446 306.259C107.166 306.884 106.864 307.5 106.529 308.103C106.36 308.415 106.182 308.719 106.002 309.015C106.002 309.015 106.002 309.015 106.002 309.015L105.148 308.494C104.839 309.003 104.509 309.492 104.165 309.966L107.448 306.254ZM107.448 306.254C108.733 303.34 109.451 300.114 109.451 296.727M107.448 306.254L109.451 296.727M109.451 296.727C109.451 283.706 98.8972 273.152 85.8758 273.152C72.8545 273.152 62.3008 283.706 62.3008 296.727C62.3008 300.239 63.069 303.583 64.4603 306.588M109.451 296.727L64.4603 306.588M64.4603 306.588C65.0881 307.955 65.8465 309.249 66.7096 310.455L64.4603 306.588ZM104.754 310.855C104.775 310.83 104.799 310.799 104.824 310.762L66.7098 310.456C66.7895 310.567 66.8677 310.672 66.9463 310.775C67.3076 311.263 67.6905 311.731 68.0763 312.181L68.0804 312.186C68.9437 313.18 69.8885 314.1 70.9051 314.93C71.6243 315.533 72.3789 316.085 73.169 316.586C75.3435 317.981 77.7655 319.028 80.3491 319.65L80.3498 319.65C82.1199 320.075 83.9754 320.302 85.8758 320.302H85.8874C86.0235 320.302 86.1913 320.302 86.3648 320.292C87.3372 320.276 88.2915 320.2 89.2321 320.065C91.933 319.68 94.4775 318.84 96.7987 317.627L96.8045 317.624C97.1251 317.453 97.4408 317.278 97.7518 317.097C98.2117 316.832 98.6574 316.548 99.0881 316.257L99.0885 316.257C99.6832 315.856 100.255 315.426 100.803 314.978L100.806 314.975C102.278 313.764 103.607 312.383 104.754 310.855Z" fill="#85E0D3" stroke="white" stroke-width="2"/>
-<path d="M100.171 314.202C99.6427 314.633 99.0955 315.045 98.529 315.427C98.1125 315.708 97.6864 315.979 97.2506 316.231C94.7907 317.669 92.0354 318.652 89.0912 319.074C88.1857 319.204 87.2656 319.277 86.3262 319.292C86.1761 319.301 86.026 319.301 85.8759 319.301C84.0552 319.301 82.278 319.083 80.5832 318.677C78.1087 318.081 75.7892 317.079 73.707 315.742C72.9516 315.263 72.2301 314.735 71.5425 314.159C70.5692 313.365 69.6637 312.483 68.8356 311.529C68.4579 311.089 68.0899 310.638 67.7461 310.173C69.286 306.527 72.0994 301.302 75.9345 300.024C81.9729 298.009 94.8391 300.847 94.8391 300.847C94.8391 300.847 99.2456 302.987 100.171 314.202Z" fill="#FC7A82"/>
-<path d="M97.251 316.231C94.7911 317.669 92.0358 318.652 89.0917 319.074C88.1861 319.204 87.2661 319.277 86.3267 319.292C86.1766 319.301 86.0265 319.301 85.8763 319.301C84.0556 319.301 82.2785 319.083 80.5837 318.677C78.1092 318.081 75.7897 317.079 73.7075 315.742C72.9521 315.263 72.2306 314.735 71.543 314.159C71.5817 308.256 72.0272 302.862 73.4557 301.573C74.1724 300.925 75.0295 300.416 75.935 300.024C78.0317 299.118 80.3754 298.813 81.9153 298.712C82.937 298.644 83.6004 298.668 83.6004 298.668C83.6004 298.668 84.9708 298.721 86.8109 298.925C87.3194 298.988 87.8327 299.051 88.3702 299.128C91.5032 299.578 95.1737 300.455 96.1228 302.111C96.9411 303.535 97.2026 309.912 97.251 316.231Z" fill="#FC7A82"/>
-<path d="M90.2107 286.844L80.8795 293.909C80.8795 293.909 75.795 288.805 77.8627 285.89C79.146 284.074 81.7415 283.202 83.5186 282.801C83.722 282.757 83.9108 282.718 84.09 282.684C84.8551 282.534 85.3684 282.491 85.3684 282.491L85.3732 282.52C85.6977 282.399 88.9324 280.06 90.7047 280.835C93.1399 281.895 90.2107 286.844 90.2107 286.844Z" fill="black"/>
-<path d="M81.8475 295.012L81.54 300.026C81.54 300.026 82.356 301.523 85.274 301.941C87.4482 302.253 88.2457 301.806 88.2457 301.806L88.0971 299.249L87.9489 296.754L82.7245 295.264L81.8475 295.012Z" fill="#FCC898"/>
-<path d="M82.7236 295.264C82.728 295.401 82.7701 295.567 82.8399 295.757C83.8078 298.33 86.8421 299.058 88.0967 299.249L87.9485 296.754L82.7236 295.264Z" fill="black"/>
-<path d="M89.7664 285.395C89.7664 285.395 92.0118 291.842 90.7605 295.271C89.5088 298.7 86.3588 297.903 84.5003 297.078C82.6423 296.254 81.8457 295.014 81.8457 295.014C81.8457 295.014 80.0148 295.482 78.949 293.73C77.8832 291.979 78.7645 290.13 79.8173 289.772C80.87 289.413 82.0249 290.242 82.0249 290.242L81.2148 288.356C81.2138 288.357 85.9182 279.451 89.7664 285.395Z" fill="#FCC898"/>
-<path d="M81.6862 292.98C81.6325 292.98 81.5778 292.962 81.5332 292.925L79.374 291.162C79.2704 291.078 79.2554 290.925 79.3396 290.821C79.4253 290.718 79.5774 290.703 79.6805 290.787L81.8397 292.55C81.9434 292.634 81.9584 292.787 81.8741 292.891C81.8257 292.95 81.7564 292.98 81.6862 292.98Z" fill="black"/>
-<path d="M79.6284 292.826C79.5151 292.826 79.4139 292.745 79.3911 292.63C79.3659 292.498 79.4516 292.372 79.5834 292.346L81.0942 292.054C81.2239 292.028 81.3518 292.114 81.3779 292.246C81.4031 292.377 81.3174 292.504 81.1862 292.529L79.6754 292.821C79.6594 292.825 79.6439 292.826 79.6284 292.826Z" fill="black"/>
-<path d="M85.4247 288.092C85.368 288.092 85.3114 288.072 85.2654 288.032C85.1646 287.944 85.1545 287.792 85.2426 287.691C85.9941 286.831 86.9161 287.055 87.2987 287.292C87.412 287.363 87.4468 287.513 87.3761 287.626C87.3045 287.739 87.1563 287.774 87.043 287.704C86.9636 287.656 86.2604 287.263 85.6062 288.01C85.5588 288.064 85.492 288.092 85.4247 288.092Z" fill="black"/>
-<path d="M90.2956 287.887C90.2366 287.887 90.177 287.865 90.13 287.821C89.6531 287.374 89.1015 287.637 89.0788 287.648C88.9592 287.708 88.8144 287.659 88.7543 287.54C88.6948 287.421 88.7417 287.277 88.8599 287.216C88.8948 287.198 89.7291 286.784 90.4612 287.467C90.5586 287.559 90.5639 287.712 90.4724 287.81C90.4249 287.861 90.3605 287.887 90.2956 287.887Z" fill="black"/>
-<path d="M89.1582 292.604C89.1306 292.604 89.1025 292.599 89.0749 292.589C88.949 292.543 88.8846 292.404 88.9306 292.279C89.5678 290.539 88.7388 289.53 88.703 289.488C88.6168 289.387 88.6289 289.234 88.7301 289.147C88.8313 289.06 88.9829 289.071 89.07 289.173C89.1136 289.223 90.1242 290.426 89.3853 292.445C89.3494 292.543 89.2569 292.604 89.1582 292.604Z" fill="black"/>
-<path d="M90.1248 290.294C90.0125 290.294 89.9123 290.216 89.8885 290.101L89.645 288.932C89.6411 288.915 89.6396 288.899 89.6396 288.882C89.6396 288.77 89.7181 288.669 89.8324 288.645C89.9602 288.618 90.0914 288.701 90.1185 288.833L90.3621 290.002C90.366 290.019 90.3675 290.035 90.3675 290.051C90.3675 290.164 90.289 290.265 90.1747 290.288C90.1578 290.292 90.1413 290.294 90.1248 290.294Z" fill="black"/>
-<path d="M86.4617 290.379C86.3498 290.379 86.2491 290.3 86.2254 290.186L85.9818 289.017C85.9542 288.886 86.0385 288.758 86.1697 288.731C86.298 288.703 86.4288 288.788 86.4559 288.918L86.6994 290.087C86.727 290.219 86.6428 290.346 86.5116 290.373C86.4951 290.377 86.4786 290.379 86.4617 290.379Z" fill="black"/>
-<path d="M90.2084 286.843C90.2084 286.843 89.7256 285.802 88.4758 285.717C87.2255 285.632 86.4415 286.468 86.4415 286.468C86.4415 286.468 86.8609 288.608 85.0915 289.622C83.3221 290.636 82.0253 290.243 82.0253 290.243L80.8186 289.728L80.0244 287.87L83.8291 284.243L88.5973 282.088L90.82 284.369L90.2084 286.843Z" fill="black"/>
-<path d="M87.7746 294.524C87.5794 294.524 86.4245 294.475 85.9742 293.339C85.9248 293.215 85.9858 293.074 86.1098 293.025C86.2337 292.975 86.3746 293.036 86.424 293.161C86.7814 294.063 87.7654 294.04 87.7847 294.04C87.9227 294.04 88.0302 294.143 88.0351 294.274C88.0394 294.407 87.9348 294.519 87.8012 294.524C87.7978 294.524 87.7891 294.524 87.7746 294.524Z" fill="black"/>
-<path d="M86.8769 302.331C83.5942 302.331 81.0961 300.057 80.975 299.945C80.8772 299.854 80.8709 299.702 80.9619 299.604C81.0534 299.505 81.2055 299.499 81.3043 299.59C81.3362 299.62 84.6063 302.593 88.4138 301.662C88.5426 301.63 88.6744 301.71 88.7063 301.84C88.7378 301.97 88.6584 302.101 88.5286 302.132C87.962 302.271 87.409 302.331 86.8769 302.331Z" fill="white"/>
-<path d="M104.166 309.966C104.103 310.053 104.045 310.136 103.982 310.218C102.878 311.695 101.595 313.031 100.171 314.203C99.648 314.629 99.1057 315.041 98.544 315.419L94.3408 313.627L95.8463 301.731L104.166 309.966Z" fill="#FC7A82"/>
-<path d="M107.298 306.949L107.296 306.955C107.015 307.58 106.713 308.195 106.379 308.799C106.21 309.111 106.032 309.415 105.852 309.71C105.852 309.71 105.852 309.711 105.852 309.711L104.998 309.19C104.688 309.698 104.359 310.187 104.015 310.662L107.298 306.949ZM107.298 306.949C108.583 304.036 109.301 300.81 109.301 297.423M107.298 306.949L109.301 297.423M109.301 297.423C109.301 284.402 98.7468 273.848 85.7255 273.848C72.7041 273.848 62.1504 284.402 62.1504 297.423C62.1504 300.935 62.9186 304.278 64.3099 307.284M109.301 297.423L64.3099 307.284M64.3099 307.284C64.9377 308.651 65.6961 309.945 66.5592 311.151L64.3099 307.284ZM104.603 311.551C104.624 311.526 104.649 311.495 104.674 311.458L66.5594 311.152C66.6391 311.263 66.7173 311.367 66.7959 311.471C67.1572 311.959 67.5401 312.427 67.9259 312.877L67.93 312.882C68.7933 313.876 69.7381 314.796 70.7548 315.626C71.4739 316.228 72.2286 316.78 73.0186 317.282C75.1931 318.677 77.6151 319.724 80.1987 320.346L80.1994 320.346C81.9695 320.771 83.825 320.998 85.7255 320.998H85.737C85.8732 320.998 86.0409 320.998 86.2144 320.988C87.1868 320.972 88.1411 320.896 89.0817 320.76C91.7826 320.376 94.3272 319.536 96.6483 318.323L96.6541 318.32C96.9747 318.149 97.2905 317.974 97.6014 317.793C98.0613 317.528 98.5071 317.244 98.9377 316.953L98.9381 316.953C99.5328 316.551 100.104 316.121 100.653 315.674L100.656 315.671C102.127 314.46 103.457 313.078 104.603 311.551Z" fill="white" stroke="white" stroke-width="2"/>
-<path d="M82.2448 301.027L78.7748 297.557C78.5879 297.37 78.3343 297.265 78.0698 297.265C77.8052 297.265 77.5516 297.37 77.3648 297.557C76.9748 297.947 76.9748 298.577 77.3648 298.967L81.5448 303.147C81.9348 303.537 82.5648 303.537 82.9548 303.147L93.5348 292.567C93.9248 292.177 93.9248 291.547 93.5348 291.157C93.3479 290.97 93.0943 290.865 92.8298 290.865C92.5652 290.865 92.3116 290.97 92.1248 291.157L82.2448 301.027Z" fill="#8E94E9"/>
-<path d="M167.53 308.374L167.53 308.374C166.973 309.35 166.347 310.285 165.66 311.17C164.65 312.473 163.508 313.665 162.249 314.727L162.246 314.73C162.221 314.751 162.174 314.792 162.118 314.834C160.597 316.105 158.916 317.185 157.105 318.046L167.53 308.374ZM167.53 308.374C169.487 304.934 170.605 300.958 170.605 296.727C170.605 283.706 160.046 273.152 147.03 273.152C134.009 273.152 123.45 283.706 123.45 296.727C123.45 301.156 124.674 305.299 126.798 308.843L126.799 308.845C127.37 309.794 128.006 310.704 128.704 311.564L167.53 308.374ZM154.565 319.073L154.563 319.074C152.197 319.867 149.664 320.302 147.03 320.302C144.771 320.302 142.587 319.984 140.517 319.387M154.565 319.073L135.854 316.344L135.357 317.211C135.357 317.212 135.358 317.212 135.358 317.212C136.967 318.133 138.697 318.866 140.517 319.387M154.565 319.073C155.435 318.78 156.285 318.436 157.105 318.046L154.565 319.073ZM140.517 319.387C140.516 319.387 140.516 319.387 140.516 319.387L140.793 318.426L140.518 319.387C140.518 319.387 140.517 319.387 140.517 319.387ZM129.455 312.442C129.456 312.443 129.457 312.444 129.458 312.445L130.198 311.772L129.452 312.438C129.453 312.44 129.454 312.441 129.455 312.442ZM129.164 312.111C129.166 312.113 129.167 312.114 129.169 312.116L129.917 311.453L129.16 312.106L129.164 312.111Z" fill="#C0A2FC" stroke="white" stroke-width="2"/>
-<path d="M143.544 287.084C143.544 287.084 139.63 283.954 141.412 281.361C143.195 278.768 147.015 281.406 147.015 281.406C147.015 281.406 148.503 279.917 150.609 281.122C152.714 282.327 152.955 285.265 152.955 285.265C152.955 285.265 155.1 285.216 155.222 287.826C155.345 290.436 152.536 292.76 152.536 292.76L143.544 287.084Z" fill="black"/>
-<path d="M161.499 314.048C160.037 315.273 158.42 316.314 156.676 317.142C155.892 317.515 155.078 317.845 154.246 318.125C151.979 318.886 149.553 319.302 147.03 319.302C144.866 319.302 142.774 318.997 140.793 318.426C139.05 317.927 137.394 317.225 135.854 316.344L135.588 301.255C141.563 297.061 147.36 298.107 147.36 298.107C147.36 298.107 154.972 297.24 158.947 300.877C159.529 301.41 160.022 302.02 160.405 302.741C161.039 303.928 161.354 308.693 161.499 314.048Z" fill="white"/>
-<path d="M150.415 294.197L150.465 299.17C150.465 299.17 149.582 300.609 146.675 300.874C144.51 301.072 143.743 300.589 143.743 300.589L144.021 298.068L144.294 295.608L149.537 294.4L150.415 294.197Z" fill="#F2AC82"/>
-<path d="M149.535 294.4C149.524 294.536 149.474 294.697 149.396 294.882C148.307 297.377 145.27 297.943 144.02 298.068L144.293 295.608L149.535 294.4Z" fill="black"/>
-<path d="M143.073 284.283C143.073 284.283 140.524 290.544 141.588 293.998C142.651 297.453 145.807 296.824 147.686 296.104C149.565 295.383 150.416 294.198 150.416 294.198C150.416 294.198 152.203 294.754 153.346 293.076C154.489 291.399 153.712 289.525 152.689 289.117C151.666 288.71 150.483 289.47 150.483 289.47L151.379 287.647C151.379 287.647 147.181 278.602 143.073 284.283Z" fill="#F2AC82"/>
-<path d="M150.69 292.198C150.616 292.198 150.543 292.164 150.495 292.099C150.416 291.991 150.439 291.84 150.547 291.76L152.772 290.127C152.879 290.049 153.031 290.07 153.11 290.179C153.189 290.287 153.166 290.438 153.059 290.517L150.834 292.151C150.79 292.183 150.741 292.198 150.69 292.198Z" fill="black"/>
-<path d="M152.732 292.15C152.713 292.15 152.693 292.147 152.674 292.142L151.194 291.777C151.064 291.745 150.985 291.614 151.017 291.484C151.048 291.355 151.176 291.275 151.31 291.308L152.79 291.673C152.92 291.705 152.999 291.836 152.967 291.966C152.94 292.076 152.841 292.15 152.732 292.15Z" fill="black"/>
-<path d="M147.261 286.909C147.234 286.909 147.206 286.904 147.178 286.894C146.811 286.761 145.755 286.702 145.378 286.69C145.245 286.686 145.14 286.575 145.144 286.441C145.148 286.307 145.25 286.205 145.393 286.206C145.54 286.211 146.838 286.255 147.344 286.439C147.47 286.485 147.534 286.624 147.489 286.749C147.453 286.848 147.36 286.909 147.261 286.909Z" fill="black"/>
-<path d="M142.435 286.723C142.318 286.723 142.215 286.638 142.196 286.519C142.176 286.387 142.266 286.262 142.398 286.242C143.055 286.138 143.254 286.148 143.287 286.15C143.42 286.161 143.519 286.279 143.508 286.412C143.497 286.544 143.381 286.644 143.25 286.633C143.237 286.631 143.058 286.627 142.474 286.72C142.46 286.722 142.448 286.723 142.435 286.723Z" fill="black"/>
-<path d="M143.701 292.255C143.597 292.255 143.502 292.188 143.469 292.084C142.848 290.068 143.885 288.617 143.93 288.557C144.008 288.449 144.159 288.425 144.267 288.504C144.375 288.583 144.399 288.734 144.321 288.842C144.311 288.855 143.384 290.165 143.931 291.941C143.971 292.069 143.899 292.204 143.772 292.244C143.748 292.251 143.725 292.255 143.701 292.255Z" fill="black"/>
-<path d="M142.482 289.112C142.462 289.112 142.442 289.11 142.421 289.104C142.291 289.07 142.214 288.938 142.248 288.808L142.548 287.665C142.582 287.535 142.713 287.456 142.844 287.492C142.973 287.526 143.05 287.658 143.016 287.788L142.716 288.931C142.688 289.04 142.59 289.112 142.482 289.112Z" fill="black"/>
-<path d="M145.63 289.142C145.609 289.142 145.589 289.14 145.568 289.134C145.439 289.101 145.361 288.968 145.396 288.839L145.696 287.695C145.729 287.565 145.86 287.487 145.991 287.522C146.12 287.556 146.198 287.688 146.164 287.818L145.864 288.961C145.836 289.071 145.737 289.142 145.63 289.142Z" fill="black"/>
-<path d="M142.562 285.693C142.562 285.693 143.093 284.688 144.333 284.667C145.574 284.646 146.307 285.513 146.307 285.513C146.307 285.513 145.784 287.608 147.481 288.701C149.179 289.794 150.482 289.471 150.482 289.471L151.701 289.022L152.581 287.225L149.003 283.446L144.398 281.072L142.084 283.215L142.562 285.693Z" fill="black"/>
-<path d="M144.657 293.416C144.597 293.416 144.56 293.414 144.551 293.413C144.419 293.401 144.319 293.284 144.331 293.151C144.343 293.017 144.47 292.924 144.592 292.93C144.638 292.933 145.585 292.996 145.98 292.131C146.035 292.009 146.18 291.958 146.301 292.012C146.422 292.067 146.476 292.211 146.42 292.333C145.965 293.33 144.982 293.416 144.657 293.416Z" fill="black"/>
-<path d="M144.02 298.067C144.02 298.067 142.994 299.155 142.383 301.161C142.036 302.299 142.121 303.978 142.36 303.98C142.599 303.982 144.39 301.211 145.532 301.042C146.674 300.874 147.63 304.191 148.118 304.307C148.605 304.423 150.502 303.118 151.175 301.228C151.848 299.338 150.64 298.12 150.64 298.12C150.64 298.12 144.745 302.595 144.02 298.067Z" fill="#92B4EA"/>
-<path d="M140.794 318.426C139.051 317.927 137.395 317.225 135.855 316.344C133.729 315.133 131.816 313.583 130.199 311.772C130.102 311.666 130.01 311.559 129.918 311.453C130.131 310.799 130.363 310.145 130.62 309.511C132.949 303.71 135.588 301.255 135.588 301.255L140.794 318.426Z" fill="white"/>
-<path d="M164.87 310.557C163.902 311.806 162.807 312.949 161.601 313.966C161.568 313.995 161.534 314.024 161.5 314.048C160.037 315.273 158.42 316.315 156.677 317.143L158.948 300.877C158.948 300.877 162.134 303.332 164.87 310.557Z" fill="white"/>
-<path d="M135.54 315.427C135.532 315.427 135.523 315.426 135.515 315.425C135.382 315.412 135.285 315.293 135.299 315.16L135.76 310.681C135.774 310.547 135.9 310.449 136.026 310.465C136.159 310.479 136.256 310.597 136.242 310.73L135.781 315.21C135.768 315.334 135.663 315.427 135.54 315.427Z" fill="#92B4EA"/>
-<path d="M198.251 318.113C200.347 319.089 202.611 319.763 204.982 320.09L204.992 320.095L205.163 320.116C206.146 320.237 207.16 320.302 208.181 320.302C214.783 320.302 220.749 317.587 225.026 313.222C225.088 313.16 225.167 313.081 225.243 313C225.307 312.935 225.365 312.871 225.422 312.804C227.795 310.265 229.609 307.197 230.678 303.797C231.382 301.564 231.756 299.187 231.756 296.727C231.756 283.706 221.202 273.152 208.181 273.152C195.159 273.152 184.605 283.706 184.605 296.727C184.605 300.593 185.536 304.24 187.19 307.463C187.563 308.195 187.978 308.909 188.434 309.597C188.618 309.884 188.811 310.158 188.99 310.413L188.996 310.421L188.999 310.426C191.239 313.565 194.232 316.134 197.718 317.86L197.718 317.86L197.725 317.863L197.729 317.865C197.894 317.945 198.072 318.031 198.251 318.113ZM205.248 320.118H205.261C205.27 320.118 205.264 320.119 205.248 320.118Z" fill="#B0E6FD" stroke="white" stroke-width="2"/>
-<path d="M224.674 312.14C224.626 312.198 224.578 312.252 224.524 312.305C224.456 312.378 224.384 312.45 224.316 312.518C220.219 316.702 214.505 319.302 208.181 319.302C207.203 319.302 206.23 319.239 205.286 319.123C205.276 319.118 205.271 319.118 205.261 319.118C202.937 318.813 200.719 318.159 198.671 317.205C198.865 309.027 198.782 301.87 198.782 301.734C198.86 301.565 202.646 300.591 207.687 300.427C215.62 300.175 222.321 304.649 224.674 312.14Z" fill="#F6FF82"/>
-<path d="M205.287 319.123C205.277 319.118 205.272 319.118 205.263 319.118C202.938 318.813 200.72 318.159 198.672 317.205C198.503 317.128 198.333 317.046 198.164 316.963C194.827 315.312 191.961 312.852 189.815 309.845C194.164 303.884 197.718 302.146 198.783 301.734C199.001 301.647 199.113 301.618 199.113 301.618L205.287 319.123Z" fill="#F6FF82"/>
-<path d="M208.699 287.287C208.699 287.287 213.82 287.661 214.233 282.751C214.673 277.523 209.842 276.9 207.732 278.965C205.622 281.03 205.949 283.192 205.949 283.192C205.949 283.192 204.548 281.864 202.944 283.412C201.767 284.548 202.311 286.468 202.311 286.468C202.311 286.468 199.669 286.3 199.858 288.615C200.047 290.931 201.669 292.144 201.669 292.144L208.699 287.287Z" fill="black"/>
-<path d="M203.783 296.224L204.18 301.686C204.18 301.686 205.279 303.184 208.493 303.212C210.887 303.233 211.684 302.634 211.684 302.634L211.152 299.892L210.629 297.218L204.767 296.367L203.783 296.224Z" fill="#A95832"/>
-<path d="M204.768 296.368C204.792 296.516 204.862 296.688 204.965 296.883C206.385 299.523 209.769 299.868 211.152 299.892L210.629 297.218L204.768 296.368Z" fill="black"/>
-<path d="M210.943 284.68C210.943 284.68 214.307 291.32 213.453 295.207C212.599 299.094 209.079 298.69 206.952 298.069C204.824 297.449 203.783 296.225 203.783 296.225C203.783 296.225 201.873 296.997 200.466 295.259C199.06 293.523 199.743 291.396 200.829 290.856C201.914 290.316 203.282 291.044 203.282 291.044L202.133 289.124C202.133 289.124 205.921 278.819 210.943 284.68Z" fill="#A95832"/>
-<path d="M203.279 294.032C203.236 294.032 203.192 294.021 203.153 293.996L201.024 292.688C200.91 292.618 200.874 292.469 200.944 292.355C201.014 292.241 201.163 292.205 201.277 292.276L203.406 293.583C203.52 293.654 203.556 293.802 203.486 293.917C203.44 293.991 203.361 294.032 203.279 294.032Z" fill="black"/>
-<path d="M201.511 294.176C201.421 294.176 201.336 294.126 201.293 294.041C201.234 293.921 201.284 293.776 201.404 293.717L202.518 293.17C202.638 293.112 202.783 293.161 202.842 293.281C202.901 293.401 202.852 293.546 202.731 293.605L201.617 294.152C201.583 294.168 201.547 294.176 201.511 294.176Z" fill="black"/>
-<path d="M211.841 287.276C211.792 287.276 211.743 287.262 211.701 287.231C211.104 286.808 210.538 287.181 210.514 287.198C210.403 287.271 210.253 287.244 210.178 287.134C210.103 287.024 210.13 286.874 210.24 286.799C210.274 286.775 211.104 286.214 211.981 286.837C212.09 286.914 212.116 287.066 212.038 287.175C211.991 287.241 211.917 287.276 211.841 287.276Z" fill="black"/>
-<path d="M210.954 293.462C210.936 293.462 210.919 293.46 210.901 293.456C210.771 293.427 210.688 293.298 210.717 293.167C211.162 291.145 210.016 289.794 210.004 289.781C209.917 289.68 209.928 289.527 210.029 289.439C210.13 289.352 210.282 289.363 210.371 289.464C210.424 289.526 211.687 291.015 211.19 293.271C211.165 293.385 211.064 293.462 210.954 293.462Z" fill="black"/>
-<path d="M211.763 290.493C211.631 290.493 211.523 290.387 211.521 290.255L211.498 288.953C211.496 288.819 211.602 288.709 211.736 288.706C211.876 288.721 211.98 288.81 211.982 288.944L212.005 290.246C212.008 290.379 211.901 290.49 211.767 290.492C211.766 290.493 211.764 290.493 211.763 290.493Z" fill="black"/>
-<path d="M208.179 290.744C208.047 290.744 207.939 290.639 207.937 290.507L207.914 289.205C207.912 289.071 208.018 288.961 208.152 288.959C208.276 288.944 208.396 289.063 208.398 289.196L208.421 290.498C208.424 290.632 208.317 290.742 208.183 290.744C208.182 290.744 208.18 290.744 208.179 290.744Z" fill="black"/>
-<path d="M211.63 286.182C211.63 286.182 210.957 285.127 209.594 285.217C208.231 285.306 207.505 286.324 207.505 286.324C207.505 286.324 208.269 288.575 206.505 289.928C204.741 291.281 203.282 291.045 203.282 291.045L201.904 290.663L200.775 288.771L204.359 284.299L209.198 281.278L211.931 283.42L211.63 286.182Z" fill="black"/>
-<path d="M209.842 294.829C209.385 294.829 208.521 294.7 208.019 293.805C207.954 293.689 207.995 293.541 208.112 293.476C208.229 293.411 208.376 293.452 208.441 293.569C208.963 294.497 210.018 294.338 210.063 294.331C210.194 294.306 210.319 294.398 210.341 294.53C210.362 294.661 210.274 294.786 210.143 294.808C210.127 294.811 210.013 294.829 209.842 294.829Z" fill="black"/>
-<path d="M213.199 292.492C213.18 292.492 213.16 292.492 213.141 292.491C211.859 292.459 210.842 291.358 210.876 290.037C210.892 289.399 211.148 288.804 211.596 288.363C212.047 287.92 212.639 287.682 213.261 287.7C214.543 287.732 215.56 288.833 215.526 290.154C215.51 290.792 215.254 291.387 214.806 291.828C214.369 292.257 213.801 292.492 213.199 292.492ZM213.204 288.183C212.73 288.183 212.282 288.368 211.936 288.708C211.577 289.06 211.373 289.537 211.36 290.05C211.333 291.103 212.138 291.981 213.153 292.007C213.647 292.025 214.11 291.833 214.466 291.482C214.825 291.13 215.029 290.653 215.042 290.141C215.068 289.087 214.264 288.209 213.248 288.183C213.234 288.183 213.219 288.183 213.204 288.183Z" fill="black"/>
-<path d="M207.675 292.886C207.655 292.886 207.635 292.886 207.614 292.886C206.974 292.87 206.379 292.605 205.938 292.141C205.497 291.678 205.262 291.07 205.278 290.43C205.294 289.79 205.558 289.194 206.022 288.753C206.486 288.312 207.096 288.081 207.734 288.094C209.055 288.127 210.103 289.229 210.07 290.55C210.037 291.851 208.968 292.886 207.675 292.886ZM207.673 288.578C207.18 288.578 206.714 288.764 206.356 289.104C205.986 289.457 205.775 289.932 205.762 290.443C205.75 290.953 205.936 291.438 206.288 291.808C206.64 292.178 207.115 292.389 207.626 292.402C208.676 292.426 209.559 291.592 209.585 290.538C209.612 289.484 208.776 288.605 207.722 288.578C207.706 288.578 207.689 288.578 207.673 288.578Z" fill="black"/>
-<path d="M209.602 290.531C209.475 290.531 209.368 290.431 209.361 290.302C209.353 290.169 209.455 290.054 209.589 290.047L211.276 289.952C211.406 289.935 211.524 290.046 211.532 290.18C211.539 290.313 211.437 290.428 211.304 290.435L209.616 290.53C209.611 290.531 209.607 290.531 209.602 290.531Z" fill="black"/>
-<path d="M205.512 290.995C205.473 290.995 205.433 290.986 205.397 290.966C205.375 290.954 203.137 289.76 202.076 290.833C201.982 290.928 201.829 290.929 201.734 290.835C201.639 290.741 201.638 290.588 201.732 290.493C203.049 289.161 205.524 290.484 205.629 290.541C205.746 290.605 205.789 290.751 205.726 290.869C205.681 290.949 205.598 290.995 205.512 290.995Z" fill="black"/>
-<path d="M208.702 303.568C207.303 303.568 205.559 303.181 203.647 301.881C203.536 301.806 203.508 301.656 203.583 301.545C203.659 301.435 203.809 301.406 203.92 301.481C208.336 304.485 211.66 302.419 211.799 302.33C211.912 302.257 212.061 302.29 212.134 302.402C212.206 302.514 212.174 302.664 212.062 302.736C212.04 302.75 210.737 303.568 208.702 303.568Z" fill="white"/>
-<path d="M206.561 287.917C206.515 287.917 206.469 287.905 206.428 287.877C206.317 287.804 206.286 287.653 206.36 287.542C207.039 286.516 208.054 286.623 208.496 286.821C208.618 286.875 208.672 287.019 208.618 287.141C208.563 287.262 208.42 287.316 208.299 287.263C208.259 287.245 207.38 286.876 206.764 287.809C206.717 287.879 206.64 287.917 206.561 287.917Z" fill="black"/>
-<defs>
-<clipPath id="clip0_1631_30438">
-<rect x="38" width="160" height="160" rx="80" fill="white"/>
-</clipPath>
-</defs>
-</svg>
-  )
-}
