+++ fwb/middleware.ts
@@ -4,7 +4,7 @@ import { authMiddleware } from "@clerk/nextjs";
 // Please edit this to allow other routes to be public as needed.
 // See https://clerk.com/docs/references/nextjs/auth-middleware for more information about configuring your Middleware
 export default authMiddleware({
-      publicRoutes: ["/","/forgotpassword",'/success', '/testing']
+      publicRoutes: ["/","/forgotpassword",'/success',]
 });
  
 export const config = {
